refactor: move NativePtyProcess and NativeTerminalInput to running-process-core
Extract PTY and terminal input infrastructure from the PyO3 bindings
crate into running-process-core so other Rust crates can use PTY
functionality without depending on PyO3.

- NativePtyProcess, IdleDetectorCore, platform modules → core/src/pty/
- TerminalInputCore, translate_console_key_event, worker → core/src/pty/terminal_input.rs
- Add portable-pty and sysinfo as regular deps to core
- Py crate wrappers delegate to core via inner field
- Bump version to 3.0.14
