+++ Cargo.toml
@@ -12,7 +12,7 @@ members = [
 ]
 
 [workspace.package]
-version = "3.0.13"
+version = "3.0.14"
 edition = "2021"
 rust-version = "1.85"
 license = "BSD-3-Clause"
