+++ ci/tiny_pdb_symbols.py
@@ -111,138 +111,7 @@ TINY_PDB_SYMBOLS: tuple[TinyPdbSymbolSpec, ...] = (
         "pub extern \"C\" fn rp_native_running_process_send_interrupt_public(",
         "api",
     ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_start_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_start_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_respond_to_queries_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_respond_to_queries_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_resize_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_resize_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_send_interrupt_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_send_interrupt_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_wait_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_wait_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_terminate_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_terminate_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_kill_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_kill_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_terminate_tree_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_terminate_tree_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_kill_tree_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_kill_tree_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_close_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_close_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_close_impl_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_close_impl_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_native_pty_process_close_nonblocking_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_native_pty_process_close_nonblocking_public(",
-        "pty",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_pty_windows_input_payload_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_pty_windows_input_payload_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_pty_windows_respond_to_queries_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_pty_windows_respond_to_queries_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_pty_windows_send_interrupt_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_pty_windows_send_interrupt_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_pty_windows_terminate_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_pty_windows_terminate_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_pty_windows_kill_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_pty_windows_kill_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_pty_windows_terminate_tree_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_pty_windows_terminate_tree_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_pty_windows_kill_tree_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_pty_windows_kill_tree_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_spawn_pty_reader_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_spawn_pty_reader_public(",
-        "bridge",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_py_assign_child_to_windows_kill_on_close_job_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_py_assign_child_to_windows_kill_on_close_job_public(",
-        "win32",
-    ),
-    TinyPdbSymbolSpec(
-        "rp_apply_windows_pty_priority_public",
-        "crates/running-process-py/src/public_symbols.rs",
-        "pub extern \"C\" fn rp_apply_windows_pty_priority_public(",
-        "win32",
-    ),
+    # PTY public symbols moved to running-process-core/src/pty/
 )
 
 DISALLOWED_PUBLIC_SYMBOL_PATTERNS: tuple[str, ...] = (
