+++ crates/running-process-py/src/lib.rs
@@ -1,20 +1,16 @@
 use std::collections::{HashMap, VecDeque};
-use std::ffi::OsString;
 #[cfg(windows)]
 use std::fs;
-#[cfg(windows)]
-use std::fs::OpenOptions;
-use std::io::{Read, Write};
 use std::path::PathBuf;
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
+use std::sync::atomic::{AtomicBool, Ordering};
+#[cfg(test)]
+use std::sync::atomic::AtomicUsize;
 use std::sync::Arc;
 use std::sync::{Condvar, Mutex, OnceLock};
 use std::thread;
 use std::time::Duration;
 use std::time::Instant;
-use std::time::{SystemTime, UNIX_EPOCH};
 
-use portable_pty::{native_pty_system, CommandBuilder, MasterPty, PtySize};
 use pyo3::exceptions::{PyRuntimeError, PyTimeoutError, PyValueError};
 use pyo3::prelude::*;
 use pyo3::types::{PyBytes, PyDict, PyList, PyString};
@@ -24,6 +20,18 @@ use running_process_core::{
     ContainedProcessGroup, Containment, NativeProcess, OriginatorProcessInfo, ProcessConfig,
     ProcessError, ReadStatus, StderrMode, StdinMode, StreamEvent, StreamKind,
 };
+use running_process_core::pty::{
+    self as core_pty, IdleDetectorCore, IdleMonitorState, NativePtyProcess as CoreNativePtyProcess,
+    PtyError,
+    terminal_input::{
+        self as core_terminal_input, TerminalInputCore, TerminalInputEventRecord,
+        TerminalInputError,
+    },
+};
+#[cfg(windows)]
+use running_process_core::pty::terminal_input::{
+    TerminalInputWaitOutcome, wait_for_terminal_input_event,
+};
 #[cfg(unix)]
 use running_process_core::{
     unix_set_priority, unix_signal_process, unix_signal_process_group, UnixSignal,
@@ -31,23 +39,17 @@ use running_process_core::{
 use sysinfo::{Pid, ProcessRefreshKind, Signal, System, UpdateKind};
 
 mod daemon_client;
-#[cfg(unix)]
-mod pty_posix;
-#[cfg(windows)]
-mod pty_windows;
 mod public_symbols;
 
-#[cfg(unix)]
-use pty_posix as pty_platform;
-
-#[cfg(windows)]
-const NATIVE_TERMINAL_INPUT_TRACE_PATH_ENV: &str =
-    "RUNNING_PROCESS_NATIVE_TERMINAL_INPUT_TRACE_PATH";
+// NATIVE_TERMINAL_INPUT_TRACE_PATH_ENV is now in running_process_core::pty::terminal_input
+#[cfg(all(windows, test))]
+use running_process_core::pty::terminal_input::NATIVE_TERMINAL_INPUT_TRACE_PATH_ENV;
 
 fn to_py_err(err: impl std::fmt::Display) -> PyErr {
     PyRuntimeError::new_err(err.to_string())
 }
 
+#[cfg(test)]
 fn is_ignorable_process_control_error(err: &std::io::Error) -> bool {
     if matches!(
         err.kind(),
@@ -121,44 +123,13 @@ fn active_process_registry() -> &'static Mutex<HashMap<u32, ActiveProcessRecord>
     ACTIVE_PROCESSES.get_or_init(|| Mutex::new(HashMap::new()))
 }
 
+// unix_now_seconds is now in running_process_core::pty::terminal_input
 fn unix_now_seconds() -> f64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|duration| duration.as_secs_f64())
-        .unwrap_or(0.0)
+    core_terminal_input::unix_now_seconds()
 }
 
-#[cfg(windows)]
-fn native_terminal_input_trace_target() -> Option<String> {
-    std::env::var(NATIVE_TERMINAL_INPUT_TRACE_PATH_ENV)
-        .ok()
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-}
-
-#[cfg(windows)]
-fn append_native_terminal_input_trace_line(line: &str) {
-    let Some(target) = native_terminal_input_trace_target() else {
-        return;
-    };
-    if target == "-" {
-        eprintln!("{line}");
-        return;
-    }
-    let Ok(mut file) = OpenOptions::new().create(true).append(true).open(&target) else {
-        return;
-    };
-    let _ = writeln!(file, "{line}");
-}
-
-#[cfg(windows)]
-fn format_terminal_input_bytes(data: &[u8]) -> String {
-    if data.is_empty() {
-        return "[]".to_string();
-    }
-    let parts: Vec<String> = data.iter().map(|byte| format!("{byte:02x}")).collect();
-    format!("[{}]", parts.join(" "))
-}
+// native_terminal_input_trace_target, append_native_terminal_input_trace_line,
+// and format_terminal_input_bytes are now in running_process_core::pty::terminal_input
 
 fn register_active_process(
     pid: u32,
@@ -357,442 +328,20 @@ fn kill_process_tree_impl(pid: u32, timeout_seconds: f64) {
     }
 }
 
-#[cfg(windows)]
-fn windows_terminal_input_payload(data: &[u8]) -> Vec<u8> {
-    let mut translated = Vec::with_capacity(data.len());
-    let mut index = 0usize;
-    while index < data.len() {
-        let current = data[index];
-        if current == b'\r' {
-            translated.push(current);
-            if index + 1 < data.len() && data[index + 1] == b'\n' {
-                translated.push(b'\n');
-                index += 2;
-                continue;
-            }
-            index += 1;
-            continue;
-        }
-        if current == b'\n' {
-            translated.push(b'\r');
-            index += 1;
-            continue;
-        }
-        translated.push(current);
-        index += 1;
-    }
-    translated
-}
+// windows_terminal_input_payload is now in running_process_core::pty
+// native_terminal_input_mode, terminal_input_modifier_parameter,
+// repeat_terminal_input_bytes, repeated_modified_sequence, repeated_tilde_sequence,
+// control_character_for_unicode, trace_translated_console_key_event,
+// translate_console_key_event, and native_terminal_input_worker
+// are now in running_process_core::pty::terminal_input
 
-#[cfg(windows)]
-fn native_terminal_input_mode(original_mode: u32) -> u32 {
-    use winapi::um::wincon::{
-        ENABLE_ECHO_INPUT, ENABLE_EXTENDED_FLAGS, ENABLE_LINE_INPUT, ENABLE_PROCESSED_INPUT,
-        ENABLE_QUICK_EDIT_MODE, ENABLE_WINDOW_INPUT,
-    };
+// repeated_modified_sequence, repeated_tilde_sequence, control_character_for_unicode
+// are now in running_process_core::pty::terminal_input
 
-    (original_mode | ENABLE_EXTENDED_FLAGS | ENABLE_WINDOW_INPUT)
-        & !(ENABLE_ECHO_INPUT | ENABLE_LINE_INPUT | ENABLE_PROCESSED_INPUT | ENABLE_QUICK_EDIT_MODE)
-}
+// trace_translated_console_key_event is now in running_process_core::pty::terminal_input
 
-#[cfg(windows)]
-fn terminal_input_modifier_parameter(shift: bool, alt: bool, ctrl: bool) -> Option<u8> {
-    let value = 1 + u8::from(shift) + (u8::from(alt) * 2) + (u8::from(ctrl) * 4);
-    (value > 1).then_some(value)
-}
-
-#[cfg(windows)]
-fn repeat_terminal_input_bytes(chunk: &[u8], repeat_count: u16) -> Vec<u8> {
-    let repeat = usize::from(repeat_count.max(1));
-    let mut output = Vec::with_capacity(chunk.len() * repeat);
-    for _ in 0..repeat {
-        output.extend_from_slice(chunk);
-    }
-    output
-}
-
-#[cfg(windows)]
-fn repeated_modified_sequence(base: &[u8], modifier: Option<u8>, repeat_count: u16) -> Vec<u8> {
-    if let Some(value) = modifier {
-        let base_text = std::str::from_utf8(base).expect("VT sequence literal must be utf-8");
-        let body = base_text
-            .strip_prefix("\x1b[")
-            .expect("VT sequence literal must start with CSI");
-        let sequence = format!("\x1b[1;{value}{body}");
-        repeat_terminal_input_bytes(sequence.as_bytes(), repeat_count)
-    } else {
-        repeat_terminal_input_bytes(base, repeat_count)
-    }
-}
-
-#[cfg(windows)]
-fn repeated_tilde_sequence(number: u8, modifier: Option<u8>, repeat_count: u16) -> Vec<u8> {
-    if let Some(value) = modifier {
-        let sequence = format!("\x1b[{number};{value}~");
-        repeat_terminal_input_bytes(sequence.as_bytes(), repeat_count)
-    } else {
-        let sequence = format!("\x1b[{number}~");
-        repeat_terminal_input_bytes(sequence.as_bytes(), repeat_count)
-    }
-}
-
-#[cfg(windows)]
-fn control_character_for_unicode(unicode: u16) -> Option<u8> {
-    let upper = char::from_u32(u32::from(unicode))?.to_ascii_uppercase();
-    match upper {
-        '@' | ' ' => Some(0x00),
-        'A'..='Z' => Some((upper as u8) - b'@'),
-        '[' => Some(0x1B),
-        '\\' => Some(0x1C),
-        ']' => Some(0x1D),
-        '^' => Some(0x1E),
-        '_' => Some(0x1F),
-        _ => None,
-    }
-}
-
-#[cfg(windows)]
-fn trace_translated_console_key_event(
-    record: &winapi::um::wincontypes::KEY_EVENT_RECORD,
-    event: TerminalInputEventRecord,
-) -> TerminalInputEventRecord {
-    append_native_terminal_input_trace_line(&format!(
-        "[{:.6}] native_terminal_input raw bKeyDown={} vk={:#06x} scan={:#06x} unicode={:#06x} control={:#010x} repeat={} translated bytes={} submit={} shift={} ctrl={} alt={}",
-        unix_now_seconds(),
-        record.bKeyDown,
-        record.wVirtualKeyCode,
-        record.wVirtualScanCode,
-        unsafe { *record.uChar.UnicodeChar() },
-        record.dwControlKeyState,
-        record.wRepeatCount.max(1),
-        format_terminal_input_bytes(&event.data),
-        event.submit,
-        event.shift,
-        event.ctrl,
-        event.alt,
-    ));
-    event
-}
-
-#[cfg(windows)]
-fn translate_console_key_event(
-    record: &winapi::um::wincontypes::KEY_EVENT_RECORD,
-) -> Option<TerminalInputEventRecord> {
-    use winapi::um::wincontypes::{
-        LEFT_ALT_PRESSED, LEFT_CTRL_PRESSED, RIGHT_ALT_PRESSED, RIGHT_CTRL_PRESSED, SHIFT_PRESSED,
-    };
-    use winapi::um::winuser::{
-        VK_BACK, VK_DELETE, VK_DOWN, VK_END, VK_ESCAPE, VK_HOME, VK_INSERT, VK_LEFT, VK_NEXT,
-        VK_PRIOR, VK_RETURN, VK_RIGHT, VK_TAB, VK_UP,
-    };
-
-    if record.bKeyDown == 0 {
-        append_native_terminal_input_trace_line(&format!(
-            "[{:.6}] native_terminal_input raw bKeyDown=0 vk={:#06x} scan={:#06x} unicode={:#06x} control={:#010x} repeat={} translated=ignored",
-            unix_now_seconds(),
-            record.wVirtualKeyCode,
-            record.wVirtualScanCode,
-            unsafe { *record.uChar.UnicodeChar() },
-            record.dwControlKeyState,
-            record.wRepeatCount,
-        ));
-        return None;
-    }
-
-    let repeat_count = record.wRepeatCount.max(1);
-    let modifiers = record.dwControlKeyState;
-    let shift = modifiers & SHIFT_PRESSED != 0;
-    let alt = modifiers & (LEFT_ALT_PRESSED | RIGHT_ALT_PRESSED) != 0;
-    let ctrl = modifiers & (LEFT_CTRL_PRESSED | RIGHT_CTRL_PRESSED) != 0;
-    let virtual_key_code = record.wVirtualKeyCode;
-    let unicode = unsafe { *record.uChar.UnicodeChar() };
-
-    // Shift+Enter: send CSI u escape sequence so downstream TUI apps
-    // (e.g. Claude Code) can distinguish Shift+Enter (newline) from
-    // plain Enter (submit).  Format: ESC [ 13 ; 2 u
-    if shift && !ctrl && !alt && virtual_key_code as i32 == VK_RETURN {
-        return Some(trace_translated_console_key_event(
-            record,
-            TerminalInputEventRecord {
-                data: repeat_terminal_input_bytes(b"\x1b[13;2u", repeat_count),
-                submit: false,
-                shift,
-                ctrl,
-                alt,
-                virtual_key_code,
-                repeat_count,
-            },
-        ));
-    }
-
-    let mut data = if ctrl {
-        control_character_for_unicode(unicode)
-            .map(|byte| repeat_terminal_input_bytes(&[byte], repeat_count))
-            .unwrap_or_default()
-    } else {
-        Vec::new()
-    };
-
-    if data.is_empty() && unicode != 0 {
-        if let Some(character) = char::from_u32(u32::from(unicode)) {
-            let text: String = std::iter::repeat_n(character, usize::from(repeat_count)).collect();
-            data = text.into_bytes();
-        }
-    }
-
-    if data.is_empty() {
-        let modifier = terminal_input_modifier_parameter(shift, alt, ctrl);
-        let sequence = match virtual_key_code as i32 {
-            VK_BACK => Some(b"\x08".as_slice()),
-            VK_TAB if shift => Some(b"\x1b[Z".as_slice()),
-            VK_TAB => Some(b"\t".as_slice()),
-            VK_RETURN => Some(b"\r".as_slice()),
-            VK_ESCAPE => Some(b"\x1b".as_slice()),
-            VK_UP => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_modified_sequence(b"\x1b[A", modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_DOWN => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_modified_sequence(b"\x1b[B", modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_RIGHT => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_modified_sequence(b"\x1b[C", modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_LEFT => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_modified_sequence(b"\x1b[D", modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_HOME => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_modified_sequence(b"\x1b[H", modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_END => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_modified_sequence(b"\x1b[F", modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_INSERT => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_tilde_sequence(2, modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_DELETE => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_tilde_sequence(3, modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_PRIOR => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_tilde_sequence(5, modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            VK_NEXT => {
-                return Some(trace_translated_console_key_event(
-                    record,
-                    TerminalInputEventRecord {
-                        data: repeated_tilde_sequence(6, modifier, repeat_count),
-                        submit: false,
-                        shift,
-                        ctrl,
-                        alt,
-                        virtual_key_code,
-                        repeat_count,
-                    },
-                ));
-            }
-            _ => None,
-        };
-        data = sequence.map(|chunk| repeat_terminal_input_bytes(chunk, repeat_count))?;
-    }
-
-    if alt && !data.starts_with(b"\x1b[") && !data.starts_with(b"\x1bO") {
-        let mut prefixed = Vec::with_capacity(data.len() + 1);
-        prefixed.push(0x1B);
-        prefixed.extend_from_slice(&data);
-        data = prefixed;
-    }
-
-    let event = TerminalInputEventRecord {
-        data,
-        submit: virtual_key_code as i32 == VK_RETURN && !shift,
-        shift,
-        ctrl,
-        alt,
-        virtual_key_code,
-        repeat_count,
-    };
-    Some(trace_translated_console_key_event(record, event))
-}
-
-#[cfg(windows)]
-fn native_terminal_input_worker(
-    input_handle: usize,
-    state: Arc<Mutex<TerminalInputState>>,
-    condvar: Arc<Condvar>,
-    stop: Arc<AtomicBool>,
-    capturing: Arc<AtomicBool>,
-) {
-    use winapi::shared::minwindef::DWORD;
-    use winapi::shared::winerror::WAIT_TIMEOUT;
-    use winapi::um::consoleapi::ReadConsoleInputW;
-    use winapi::um::synchapi::WaitForSingleObject;
-    use winapi::um::winbase::WAIT_OBJECT_0;
-    use winapi::um::wincontypes::{INPUT_RECORD, KEY_EVENT};
-    use winapi::um::winnt::HANDLE;
-
-    let handle = input_handle as HANDLE;
-    let mut records: [INPUT_RECORD; 512] = unsafe { std::mem::zeroed() };
-    append_native_terminal_input_trace_line(&format!(
-        "[{:.6}] native_terminal_input worker_start handle={input_handle}",
-        unix_now_seconds(),
-    ));
-
-    while !stop.load(Ordering::Acquire) {
-        let wait_result = unsafe { WaitForSingleObject(handle, 50) };
-        match wait_result {
-            WAIT_OBJECT_0 => {
-                let mut read_count: DWORD = 0;
-                let ok = unsafe {
-                    ReadConsoleInputW(
-                        handle,
-                        records.as_mut_ptr(),
-                        records.len() as DWORD,
-                        &mut read_count,
-                    )
-                };
-                if ok == 0 {
-                    append_native_terminal_input_trace_line(&format!(
-                        "[{:.6}] native_terminal_input read_console_input_failed handle={input_handle}",
-                        unix_now_seconds(),
-                    ));
-                    break;
-                }
-                let mut batch = Vec::new();
-                for record in records.iter().take(read_count as usize) {
-                    if record.EventType != KEY_EVENT {
-                        continue;
-                    }
-                    let key_event = unsafe { record.Event.KeyEvent() };
-                    if let Some(event) = translate_console_key_event(key_event) {
-                        batch.push(event);
-                    }
-                }
-                if !batch.is_empty() {
-                    let mut guard = state.lock().expect("terminal input mutex poisoned");
-                    guard.events.extend(batch);
-                    drop(guard);
-                    condvar.notify_all();
-                }
-            }
-            WAIT_TIMEOUT => continue,
-            _ => {
-                append_native_terminal_input_trace_line(&format!(
-                    "[{:.6}] native_terminal_input wait_result={wait_result} handle={input_handle}",
-                    unix_now_seconds(),
-                ));
-                break;
-            }
-        }
-    }
-
-    capturing.store(false, Ordering::Release);
-    let mut guard = state.lock().expect("terminal input mutex poisoned");
-    guard.closed = true;
-    condvar.notify_all();
-    drop(guard);
-    append_native_terminal_input_trace_line(&format!(
-        "[{:.6}] native_terminal_input worker_stop handle={input_handle}",
-        unix_now_seconds(),
-    ));
-}
+// translate_console_key_event and native_terminal_input_worker
+// are now in running_process_core::pty::terminal_input
 
 #[pyfunction]
 fn native_get_process_tree_info(pid: u32) -> String {
@@ -968,7 +517,7 @@ fn windows_generate_console_ctrl_break_impl(pid: u32, creationflags: Option<u32>
 #[pyfunction]
 fn native_windows_terminal_input_bytes(py: Python<'_>, data: &[u8]) -> Py<PyAny> {
     #[cfg(windows)]
-    let payload = windows_terminal_input_payload(data);
+    let payload = core_pty::windows_terminal_input_payload(data);
     #[cfg(not(windows))]
     let payload = data.to_vec();
     PyBytes::new(py, &payload).into_any().unbind()
@@ -1090,23 +639,7 @@ impl NativeProcessMetrics {
     }
 }
 
-struct PtyReadState {
-    chunks: VecDeque<Vec<u8>>,
-    closed: bool,
-}
-
-struct PtyReadShared {
-    state: Mutex<PtyReadState>,
-    condvar: Condvar,
-}
-
-struct NativePtyHandles {
-    master: Box<dyn MasterPty + Send>,
-    writer: Box<dyn Write + Send>,
-    child: Box<dyn portable_pty::Child + Send + Sync>,
-    #[cfg(windows)]
-    _job: WindowsJobHandle,
-}
+// PTY types are now in running_process_core::pty
 
 #[pyclass]
 struct NativeProcessMetrics {
@@ -1116,95 +649,31 @@ struct NativeProcessMetrics {
 
 #[pyclass]
 struct NativePtyProcess {
-    argv: Vec<String>,
-    cwd: Option<String>,
-    env: Option<Vec<(String, String)>>,
-    rows: u16,
-    cols: u16,
-    #[cfg(windows)]
-    nice: Option<i32>,
-    handles: Arc<Mutex<Option<NativePtyHandles>>>,
-    reader: Arc<PtyReadShared>,
-    returncode: Arc<Mutex<Option<i32>>>,
-    input_bytes_total: Arc<AtomicUsize>,
-    newline_events_total: Arc<AtomicUsize>,
-    submit_events_total: Arc<AtomicUsize>,
-    /// When true, the reader thread writes PTY output to stdout.
-    echo: Arc<AtomicBool>,
-    /// When set, the reader thread feeds output directly to the idle detector.
-    idle_detector: Arc<Mutex<Option<Arc<IdleDetectorCore>>>>,
-    /// Visible (non-control) output bytes seen by the reader thread.
-    output_bytes_total: Arc<AtomicUsize>,
-    /// Control churn bytes (ANSI escapes, BS, CR, DEL) seen by the reader.
-    control_churn_bytes_total: Arc<AtomicUsize>,
-    #[cfg(windows)]
-    terminal_input_relay_stop: Arc<AtomicBool>,
-    #[cfg(windows)]
-    terminal_input_relay_active: Arc<AtomicBool>,
-    #[cfg(windows)]
-    terminal_input_relay_worker: Mutex<Option<thread::JoinHandle<()>>>,
+    inner: CoreNativePtyProcess,
 }
 
 impl NativePtyProcess {
-    fn mark_reader_closed(&self) {
-        let mut guard = self.reader.state.lock().expect("pty read mutex poisoned");
-        guard.closed = true;
-        self.reader.condvar.notify_all();
-    }
-
-    fn store_returncode(&self, code: i32) {
-        store_pty_returncode(&self.returncode, code);
-    }
-
-    fn record_input_metrics(&self, data: &[u8], submit: bool) {
-        record_pty_input_metrics(
-            &self.input_bytes_total,
-            &self.newline_events_total,
-            &self.submit_events_total,
-            data,
-            submit,
-        );
-    }
-
-    fn write_impl(&self, data: &[u8], submit: bool) -> PyResult<()> {
-        self.record_input_metrics(data, submit);
-        write_pty_input(&self.handles, data).map_err(to_py_err)
-    }
-
-    #[cfg(windows)]
-    fn request_terminal_input_relay_stop(&self) {
-        self.terminal_input_relay_stop
-            .store(true, Ordering::Release);
-        self.terminal_input_relay_active
-            .store(false, Ordering::Release);
-    }
-
-    #[cfg(windows)]
-    fn stop_terminal_input_relay_impl(&self) {
-        self.request_terminal_input_relay_stop();
-        if let Some(worker) = self
-            .terminal_input_relay_worker
-            .lock()
-            .expect("pty terminal input relay mutex poisoned")
-            .take()
-        {
-            let _ = worker.join();
+    fn pty_err_to_py(err: PtyError) -> PyErr {
+        match err {
+            PtyError::Timeout => PyTimeoutError::new_err(err.to_string()),
+            _ => PyRuntimeError::new_err(err.to_string()),
         }
     }
 
-    #[cfg(not(windows))]
-    fn stop_terminal_input_relay_impl(&self) {}
-
     #[cfg(windows)]
-    fn start_terminal_input_relay_impl(&self) -> PyResult<()> {
+    fn start_terminal_input_relay_py(&self) -> PyResult<()> {
         let mut worker_guard = self
+            .inner
             .terminal_input_relay_worker
             .lock()
             .expect("pty terminal input relay mutex poisoned");
-        if worker_guard.is_some() && self.terminal_input_relay_active.load(Ordering::Acquire) {
+        if worker_guard.is_some()
+            && self.inner.terminal_input_relay_active.load(Ordering::Acquire)
+        {
             return Ok(());
         }
         if self
+            .inner
             .handles
             .lock()
             .expect("pty handles mutex poisoned")
@@ -1215,28 +684,30 @@ impl NativePtyProcess {
             ));
         }
 
-        let capture = NativeTerminalInput::new();
-        capture.start_impl()?;
+        let capture = TerminalInputCore::new();
+        capture.start_impl().map_err(to_py_err)?;
 
-        self.terminal_input_relay_stop
+        self.inner
+            .terminal_input_relay_stop
             .store(false, Ordering::Release);
-        self.terminal_input_relay_active
+        self.inner
+            .terminal_input_relay_active
             .store(true, Ordering::Release);
 
-        let handles = Arc::clone(&self.handles);
-        let returncode = Arc::clone(&self.returncode);
-        let input_bytes_total = Arc::clone(&self.input_bytes_total);
-        let newline_events_total = Arc::clone(&self.newline_events_total);
-        let submit_events_total = Arc::clone(&self.submit_events_total);
-        let stop = Arc::clone(&self.terminal_input_relay_stop);
-        let active = Arc::clone(&self.terminal_input_relay_active);
+        let handles = Arc::clone(&self.inner.handles);
+        let returncode = Arc::clone(&self.inner.returncode);
+        let input_bytes_total = Arc::clone(&self.inner.input_bytes_total);
+        let newline_events_total = Arc::clone(&self.inner.newline_events_total);
+        let submit_events_total = Arc::clone(&self.inner.submit_events_total);
+        let stop = Arc::clone(&self.inner.terminal_input_relay_stop);
+        let active = Arc::clone(&self.inner.terminal_input_relay_active);
 
         *worker_guard = Some(thread::spawn(move || {
             loop {
                 if stop.load(Ordering::Acquire) {
                     break;
                 }
-                match poll_pty_process(&handles, &returncode) {
+                match core_pty::poll_pty_process(&handles, &returncode) {
                     Ok(Some(_)) => break,
                     Ok(None) => {}
                     Err(_) => break,
@@ -1247,14 +718,14 @@ impl NativePtyProcess {
                     Some(Duration::from_millis(50)),
                 ) {
                     TerminalInputWaitOutcome::Event(event) => {
-                        record_pty_input_metrics(
+                        core_pty::record_pty_input_metrics(
                             &input_bytes_total,
                             &newline_events_total,
                             &submit_events_total,
                             &event.data,
                             event.submit,
                         );
-                        if write_pty_input(&handles, &event.data).is_err() {
+                        if core_pty::write_pty_input(&handles, &event.data).is_err() {
                             break;
                         }
                     }
@@ -1270,337 +741,14 @@ impl NativePtyProcess {
     }
 
     #[cfg(not(windows))]
-    fn start_terminal_input_relay_impl(&self) -> PyResult<()> {
+    fn start_terminal_input_relay_py(&self) -> PyResult<()> {
         Err(PyRuntimeError::new_err(
             "Native PTY terminal input relay is only available on Windows consoles",
         ))
     }
-
-    /// Synchronously tear down the PTY and reap the child.
-    ///
-    /// This MUST NOT be called while holding the Python GIL — `child.wait()`
-    /// can block indefinitely on Windows ConPTY (the child stays alive until
-    /// every handle to the master pipe is dropped, including the one held by
-    /// the background reader thread). Always wrap this in `py.allow_threads`
-    /// from a Python-callable method.
-    // Preserve a stable Rust frame here in release user dumps.
-    #[inline(never)]
-    fn close_impl(&self) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!(
-            "running_process_py::NativePtyProcess::close_impl"
-        );
-        self.stop_terminal_input_relay_impl();
-        let mut guard = self.handles.lock().expect("pty handles mutex poisoned");
-        let Some(handles) = guard.take() else {
-            self.mark_reader_closed();
-            return Ok(());
-        };
-        // Release the lock while we wait so other threads can still touch
-        // unrelated fields on this object (e.g. the reader buffer).
-        drop(guard);
-
-        #[cfg(windows)]
-        let NativePtyHandles {
-            master,
-            writer,
-            mut child,
-            _job,
-        } = handles;
-        #[cfg(not(windows))]
-        let NativePtyHandles {
-            master,
-            writer,
-            mut child,
-        } = handles;
-
-        // Kill first so the child has stopped writing before we tear down
-        // ConPTY. On Windows, ClosePseudoConsole (triggered by dropping
-        // master) does not always terminate the child, so we explicitly
-        // TerminateProcess it.
-        if let Err(err) = child.kill() {
-            if !is_ignorable_process_control_error(&err) {
-                return Err(to_py_err(err));
-            }
-        }
-
-        // Drop the writer/master so the background reader thread sees EOF
-        // and releases its handle. Otherwise the reader stays blocked
-        // forever holding a master clone, which keeps ConPTY alive.
-        drop(writer);
-        drop(master);
-
-        // Now block until the child is reaped. This is safe to call
-        // unbounded because `close()` invoked us inside `py.allow_threads`,
-        // so the GIL is released and other Python threads can make
-        // progress. After the explicit kill() above, this returns within
-        // milliseconds in practice.
-        let code = match child.wait() {
-            Ok(status) => portable_exit_code(status),
-            Err(_) => -9,
-        };
-        drop(child);
-        #[cfg(windows)]
-        drop(_job);
-
-        self.store_returncode(code);
-        self.mark_reader_closed();
-        Ok(())
-    }
-
-    /// Best-effort, non-blocking teardown for use from `Drop`.
-    ///
-    /// `Drop` runs while Python holds the GIL (it is invoked by PyO3 during
-    /// finalization), so we cannot call any blocking syscalls here. We kill
-    /// the child, drop every handle so the OS reclaims the file descriptors,
-    /// and let the OS reap the process. The background reader thread will
-    /// notice EOF on its master clone and exit on its own.
-    // Preserve a stable Rust frame here in release user dumps.
-    #[inline(never)]
-    fn close_nonblocking(&self) {
-        running_process_core::rp_rust_debug_scope!(
-            "running_process_py::NativePtyProcess::close_nonblocking"
-        );
-        #[cfg(windows)]
-        self.request_terminal_input_relay_stop();
-        let Ok(mut guard) = self.handles.lock() else {
-            return;
-        };
-        let Some(handles) = guard.take() else {
-            self.mark_reader_closed();
-            return;
-        };
-        drop(guard);
-
-        #[cfg(windows)]
-        let NativePtyHandles {
-            master,
-            writer,
-            mut child,
-            _job,
-        } = handles;
-        #[cfg(not(windows))]
-        let NativePtyHandles {
-            master,
-            writer,
-            mut child,
-        } = handles;
-
-        if let Err(err) = child.kill() {
-            if !is_ignorable_process_control_error(&err) {
-                return;
-            }
-        }
-        // Drop writer + master so the reader thread sees EOF immediately.
-        drop(writer);
-        drop(master);
-        // Do NOT call child.wait() here — that would block the interpreter.
-        // Drop on the child closes its OS handle; the process is reaped by
-        // the OS once it actually exits.
-        drop(child);
-        #[cfg(windows)]
-        drop(_job);
-        self.mark_reader_closed();
-    }
-
-    fn start_impl(&self) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!("running_process_py::NativePtyProcess::start");
-        let mut guard = self.handles.lock().expect("pty handles mutex poisoned");
-        if guard.is_some() {
-            return Err(PyRuntimeError::new_err(
-                "Pseudo-terminal process already started",
-            ));
-        }
-
-        let pty_system = native_pty_system();
-        let pair = pty_system
-            .openpty(PtySize {
-                rows: self.rows,
-                cols: self.cols,
-                pixel_width: 0,
-                pixel_height: 0,
-            })
-            .map_err(to_py_err)?;
-
-        let mut cmd = command_builder_from_argv(&self.argv);
-        if let Some(cwd) = &self.cwd {
-            cmd.cwd(cwd);
-        }
-        if let Some(env) = &self.env {
-            cmd.env_clear();
-            for (key, value) in env {
-                cmd.env(key, value);
-            }
-        }
-
-        let reader = pair.master.try_clone_reader().map_err(to_py_err)?;
-        let writer = pair.master.take_writer().map_err(to_py_err)?;
-        let child = pair.slave.spawn_command(cmd).map_err(to_py_err)?;
-        #[cfg(windows)]
-        let job = public_symbols::rp_py_assign_child_to_windows_kill_on_close_job_public(
-            child.as_raw_handle(),
-        )?;
-        #[cfg(windows)]
-        public_symbols::rp_apply_windows_pty_priority_public(child.as_raw_handle(), self.nice)?;
-        let shared = Arc::clone(&self.reader);
-        let echo = Arc::clone(&self.echo);
-        let idle_detector = Arc::clone(&self.idle_detector);
-        let output_bytes = Arc::clone(&self.output_bytes_total);
-        let churn_bytes = Arc::clone(&self.control_churn_bytes_total);
-        thread::spawn(move || {
-            spawn_pty_reader(
-                reader,
-                shared,
-                echo,
-                idle_detector,
-                output_bytes,
-                churn_bytes,
-            );
-        });
-
-        *guard = Some(NativePtyHandles {
-            master: pair.master,
-            writer,
-            child,
-            #[cfg(windows)]
-            _job: job,
-        });
-        Ok(())
-    }
-
-    fn respond_to_queries_impl(&self, data: &[u8]) -> PyResult<()> {
-        #[cfg(windows)]
-        {
-            public_symbols::rp_pty_windows_respond_to_queries_public(self, data)
-        }
-
-        #[cfg(unix)]
-        {
-            pty_platform::respond_to_queries(self, data)
-        }
-    }
-
-    fn resize_impl(&self, rows: u16, cols: u16) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!("running_process_py::NativePtyProcess::resize");
-        let guard = self.handles.lock().expect("pty handles mutex poisoned");
-        if let Some(handles) = guard.as_ref() {
-            handles
-                .master
-                .resize(PtySize {
-                    rows,
-                    cols,
-                    pixel_width: 0,
-                    pixel_height: 0,
-                })
-                .map_err(to_py_err)?;
-        }
-        Ok(())
-    }
-
-    fn send_interrupt_impl(&self) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!(
-            "running_process_py::NativePtyProcess::send_interrupt"
-        );
-        #[cfg(windows)]
-        {
-            public_symbols::rp_pty_windows_send_interrupt_public(self)
-        }
-
-        #[cfg(unix)]
-        {
-            pty_platform::send_interrupt(self)
-        }
-    }
-
-    fn wait_impl(&self, timeout: Option<f64>) -> PyResult<i32> {
-        running_process_core::rp_rust_debug_scope!("running_process_py::NativePtyProcess::wait");
-        // Fast path: already exited.
-        if let Some(code) = *self.returncode.lock().expect("pty returncode mutex poisoned") {
-            return Ok(code);
-        }
-        let start = Instant::now();
-        loop {
-            if let Some(code) = poll_pty_process(&self.handles, &self.returncode)
-                .map_err(to_py_err)?
-            {
-                return Ok(code);
-            }
-            if timeout.is_some_and(|limit| start.elapsed() >= Duration::from_secs_f64(limit)) {
-                return Err(PyTimeoutError::new_err("Pseudo-terminal process timed out"));
-            }
-            thread::sleep(Duration::from_millis(10));
-        }
-    }
-
-    fn terminate_impl(&self) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!(
-            "running_process_py::NativePtyProcess::terminate"
-        );
-        #[cfg(windows)]
-        {
-            public_symbols::rp_pty_windows_terminate_public(self)
-        }
-
-        #[cfg(unix)]
-        {
-            pty_platform::terminate(self)
-        }
-    }
-
-    fn kill_impl(&self) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!("running_process_py::NativePtyProcess::kill");
-        #[cfg(windows)]
-        {
-            public_symbols::rp_pty_windows_kill_public(self)
-        }
-
-        #[cfg(unix)]
-        {
-            pty_platform::kill(self)
-        }
-    }
-
-    fn terminate_tree_impl(&self) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!(
-            "running_process_py::NativePtyProcess::terminate_tree"
-        );
-        #[cfg(windows)]
-        {
-            public_symbols::rp_pty_windows_terminate_tree_public(self)
-        }
-
-        #[cfg(unix)]
-        {
-            pty_platform::terminate_tree(self)
-        }
-    }
-
-    fn kill_tree_impl(&self) -> PyResult<()> {
-        running_process_core::rp_rust_debug_scope!(
-            "running_process_py::NativePtyProcess::kill_tree"
-        );
-        #[cfg(windows)]
-        {
-            public_symbols::rp_pty_windows_kill_tree_public(self)
-        }
-
-        #[cfg(unix)]
-        {
-            pty_platform::kill_tree(self)
-        }
-    }
 }
 
-#[cfg(windows)]
-struct WindowsJobHandle(usize);
-
-#[cfg(windows)]
-impl Drop for WindowsJobHandle {
-    fn drop(&mut self) {
-        unsafe {
-            winapi::um::handleapi::CloseHandle(self.0 as winapi::shared::ntdef::HANDLE);
-        }
-    }
-}
+// WindowsJobHandle is now in running_process_core::pty
 
 fn parse_command(command: &Bound<'_, PyAny>, shell: bool) -> PyResult<CommandSpec> {
     if let Ok(command) = command.extract::<String>() {
@@ -1686,125 +834,9 @@ struct NativeSignalBool {
     write_lock: Arc<Mutex<()>>,
 }
 
-struct IdleMonitorState {
-    last_reset_at: Instant,
-    returncode: Option<i32>,
-    interrupted: bool,
-}
+// IdleMonitorState and IdleDetectorCore are now in running_process_core::pty
 
-/// Core idle detection logic, shareable across threads via Arc.
-/// The reader thread calls `record_output` directly without the GIL.
-struct IdleDetectorCore {
-    timeout_seconds: f64,
-    stability_window_seconds: f64,
-    sample_interval_seconds: f64,
-    reset_on_input: bool,
-    reset_on_output: bool,
-    count_control_churn_as_output: bool,
-    enabled: Arc<AtomicBool>,
-    state: Mutex<IdleMonitorState>,
-    condvar: Condvar,
-}
-
-impl IdleDetectorCore {
-    fn record_input(&self, byte_count: usize) {
-        if !self.reset_on_input || byte_count == 0 {
-            return;
-        }
-        let mut guard = self.state.lock().expect("idle monitor mutex poisoned");
-        guard.last_reset_at = Instant::now();
-        self.condvar.notify_all();
-    }
-
-    fn record_output(&self, data: &[u8]) {
-        if !self.reset_on_output || data.is_empty() {
-            return;
-        }
-        let control_bytes = control_churn_bytes(data);
-        let visible_output_bytes = data.len().saturating_sub(control_bytes);
-        let active_output =
-            visible_output_bytes > 0 || (self.count_control_churn_as_output && control_bytes > 0);
-        if !active_output {
-            return;
-        }
-        let mut guard = self.state.lock().expect("idle monitor mutex poisoned");
-        guard.last_reset_at = Instant::now();
-        self.condvar.notify_all();
-    }
-
-    fn mark_exit(&self, returncode: i32, interrupted: bool) {
-        let mut guard = self.state.lock().expect("idle monitor mutex poisoned");
-        guard.returncode = Some(returncode);
-        guard.interrupted = interrupted;
-        self.condvar.notify_all();
-    }
-
-    fn enabled(&self) -> bool {
-        self.enabled.load(Ordering::Acquire)
-    }
-
-    fn set_enabled(&self, enabled: bool) {
-        let was_enabled = self.enabled.swap(enabled, Ordering::AcqRel);
-        if enabled && !was_enabled {
-            let mut guard = self.state.lock().expect("idle monitor mutex poisoned");
-            guard.last_reset_at = Instant::now();
-        }
-        self.condvar.notify_all();
-    }
-
-    fn wait(&self, timeout: Option<f64>) -> (bool, String, f64, Option<i32>) {
-        let started = Instant::now();
-        let overall_timeout = timeout.map(Duration::from_secs_f64);
-        let min_idle = self.timeout_seconds.max(self.stability_window_seconds);
-        let sample_interval = Duration::from_secs_f64(self.sample_interval_seconds.max(0.001));
-
-        let mut guard = self.state.lock().expect("idle monitor mutex poisoned");
-        loop {
-            let now = Instant::now();
-            let idle_for = now.duration_since(guard.last_reset_at).as_secs_f64();
-
-            if let Some(returncode) = guard.returncode {
-                let reason = if guard.interrupted {
-                    "interrupt"
-                } else {
-                    "process_exit"
-                };
-                return (false, reason.to_string(), idle_for, Some(returncode));
-            }
-
-            let enabled = self.enabled.load(Ordering::Acquire);
-            if enabled && idle_for >= min_idle {
-                return (true, "idle_timeout".to_string(), idle_for, None);
-            }
-
-            if let Some(limit) = overall_timeout {
-                if now.duration_since(started) >= limit {
-                    return (false, "timeout".to_string(), idle_for, None);
-                }
-            }
-
-            let idle_remaining = if enabled {
-                (min_idle - idle_for).max(0.0)
-            } else {
-                sample_interval.as_secs_f64()
-            };
-            let mut wait_for =
-                sample_interval.min(Duration::from_secs_f64(idle_remaining.max(0.001)));
-            if let Some(limit) = overall_timeout {
-                let elapsed = now.duration_since(started);
-                if elapsed < limit {
-                    let remaining = limit - elapsed;
-                    wait_for = wait_for.min(remaining);
-                }
-            }
-            let result = self
-                .condvar
-                .wait_timeout(guard, wait_for)
-                .expect("idle monitor mutex poisoned");
-            guard = result.0;
-        }
-    }
-}
+// IdleDetectorCore impl is now in running_process_core::pty
 
 #[pyclass]
 struct NativeIdleDetector {
@@ -1827,129 +859,13 @@ struct NativePtyBuffer {
     condvar: Condvar,
 }
 
-#[derive(Clone)]
-struct TerminalInputEventRecord {
-    data: Vec<u8>,
-    submit: bool,
-    shift: bool,
-    ctrl: bool,
-    alt: bool,
-    virtual_key_code: u16,
-    repeat_count: u16,
-}
+// TerminalInputEventRecord, TerminalInputState, ActiveTerminalInputCapture,
+// TerminalInputWaitOutcome, and wait_for_terminal_input_event
+// are now in running_process_core::pty::terminal_input
 
-struct TerminalInputState {
-    events: VecDeque<TerminalInputEventRecord>,
-    closed: bool,
-}
-
-#[cfg(windows)]
-struct ActiveTerminalInputCapture {
-    input_handle: usize,
-    original_mode: u32,
-    active_mode: u32,
-}
-
-#[cfg(windows)]
-enum TerminalInputWaitOutcome {
-    Event(TerminalInputEventRecord),
-    Closed,
-    Timeout,
-}
-
-#[cfg(windows)]
-fn wait_for_terminal_input_event(
-    state: &Arc<Mutex<TerminalInputState>>,
-    condvar: &Arc<Condvar>,
-    timeout: Option<Duration>,
-) -> TerminalInputWaitOutcome {
-    let deadline = timeout.map(|limit| Instant::now() + limit);
-    let mut guard = state.lock().expect("terminal input mutex poisoned");
-    loop {
-        if let Some(event) = guard.events.pop_front() {
-            return TerminalInputWaitOutcome::Event(event);
-        }
-        if guard.closed {
-            return TerminalInputWaitOutcome::Closed;
-        }
-        match deadline {
-            Some(deadline) => {
-                let now = Instant::now();
-                if now >= deadline {
-                    return TerminalInputWaitOutcome::Timeout;
-                }
-                let wait = deadline.saturating_duration_since(now);
-                let result = condvar
-                    .wait_timeout(guard, wait)
-                    .expect("terminal input mutex poisoned");
-                guard = result.0;
-            }
-            None => {
-                guard = condvar.wait(guard).expect("terminal input mutex poisoned");
-            }
-        }
-    }
-}
-
-fn input_contains_newline(data: &[u8]) -> bool {
-    data.iter().any(|byte| matches!(*byte, b'\r' | b'\n'))
-}
-
-fn record_pty_input_metrics(
-    input_bytes_total: &Arc<AtomicUsize>,
-    newline_events_total: &Arc<AtomicUsize>,
-    submit_events_total: &Arc<AtomicUsize>,
-    data: &[u8],
-    submit: bool,
-) {
-    input_bytes_total.fetch_add(data.len(), Ordering::AcqRel);
-    if input_contains_newline(data) {
-        newline_events_total.fetch_add(1, Ordering::AcqRel);
-    }
-    if submit {
-        submit_events_total.fetch_add(1, Ordering::AcqRel);
-    }
-}
-
-fn store_pty_returncode(returncode: &Arc<Mutex<Option<i32>>>, code: i32) {
-    *returncode.lock().expect("pty returncode mutex poisoned") = Some(code);
-}
-
-fn poll_pty_process(
-    handles: &Arc<Mutex<Option<NativePtyHandles>>>,
-    returncode: &Arc<Mutex<Option<i32>>>,
-) -> Result<Option<i32>, std::io::Error> {
-    let mut guard = handles.lock().expect("pty handles mutex poisoned");
-    let Some(handles) = guard.as_mut() else {
-        return Ok(*returncode.lock().expect("pty returncode mutex poisoned"));
-    };
-    let status = handles.child.try_wait()?;
-    let code = status.map(portable_exit_code);
-    if let Some(code) = code {
-        store_pty_returncode(returncode, code);
-        return Ok(Some(code));
-    }
-    Ok(None)
-}
-
-fn write_pty_input(
-    handles: &Arc<Mutex<Option<NativePtyHandles>>>,
-    data: &[u8],
-) -> Result<(), std::io::Error> {
-    let mut guard = handles.lock().expect("pty handles mutex poisoned");
-    let handles = guard.as_mut().ok_or_else(|| {
-        std::io::Error::new(
-            std::io::ErrorKind::NotConnected,
-            "Pseudo-terminal process is not running",
-        )
-    })?;
-    #[cfg(windows)]
-    let payload = public_symbols::rp_pty_windows_input_payload_public(data);
-    #[cfg(unix)]
-    let payload = pty_platform::input_payload(data);
-    handles.writer.write_all(&payload)?;
-    handles.writer.flush()
-}
+// PTY helper functions (input_contains_newline, record_pty_input_metrics,
+// store_pty_returncode, poll_pty_process, write_pty_input) are now in
+// running_process_core::pty
 
 #[pyclass]
 #[derive(Clone)]
@@ -1965,13 +881,7 @@ struct NativeTerminalInputEvent {
 
 #[pyclass]
 struct NativeTerminalInput {
-    state: Arc<Mutex<TerminalInputState>>,
-    condvar: Arc<Condvar>,
-    stop: Arc<AtomicBool>,
-    capturing: Arc<AtomicBool>,
-    worker: Mutex<Option<thread::JoinHandle<()>>>,
-    #[cfg(windows)]
-    console: Mutex<Option<ActiveTerminalInputCapture>>,
+    inner: TerminalInputCore,
 }
 
 #[pymethods]
@@ -2342,10 +1252,18 @@ impl PyNativeProcess {
         cols: u16,
         nice: Option<i32>,
     ) -> PyResult<Self> {
+        let env_pairs = env
+            .map(|mapping| {
+                mapping
+                    .iter()
+                    .map(|(key, value)| Ok((key.extract::<String>()?, value.extract::<String>()?)))
+                    .collect::<PyResult<Vec<(String, String)>>>()
+            })
+            .transpose()?;
+        let inner = CoreNativePtyProcess::new(argv, cwd, env_pairs, rows, cols, nice)
+            .map_err(NativePtyProcess::pty_err_to_py)?;
         Ok(Self {
-            backend: NativeProcessBackend::Pty(NativePtyProcess::new(
-                argv, cwd, env, rows, cols, nice,
-            )?),
+            backend: NativeProcessBackend::Pty(NativePtyProcess { inner }),
         })
     }
 
@@ -2661,6 +1579,7 @@ impl PyNativeProcess {
         match &self.backend {
             NativeProcessBackend::Running(process) => Ok(process.returncode()),
             NativeProcessBackend::Pty(process) => Ok(*process
+                .inner
                 .returncode
                 .lock()
                 .expect("pty returncode mutex poisoned")),
@@ -2702,11 +1621,6 @@ impl NativePtyProcess {
         cols: u16,
         nice: Option<i32>,
     ) -> PyResult<Self> {
-        if argv.is_empty() {
-            return Err(PyValueError::new_err("command cannot be empty"));
-        }
-        #[cfg(not(windows))]
-        let _ = nice;
         let env_pairs = env
             .map(|mapping| {
                 mapping
@@ -2715,200 +1629,95 @@ impl NativePtyProcess {
                     .collect::<PyResult<Vec<(String, String)>>>()
             })
             .transpose()?;
-        Ok(Self {
-            argv,
-            cwd,
-            env: env_pairs,
-            rows,
-            cols,
-            #[cfg(windows)]
-            nice,
-            handles: Arc::new(Mutex::new(None)),
-            reader: Arc::new(PtyReadShared {
-                state: Mutex::new(PtyReadState {
-                    chunks: VecDeque::new(),
-                    closed: false,
-                }),
-                condvar: Condvar::new(),
-            }),
-            returncode: Arc::new(Mutex::new(None)),
-            input_bytes_total: Arc::new(AtomicUsize::new(0)),
-            newline_events_total: Arc::new(AtomicUsize::new(0)),
-            submit_events_total: Arc::new(AtomicUsize::new(0)),
-            echo: Arc::new(AtomicBool::new(false)),
-            idle_detector: Arc::new(Mutex::new(None)),
-            output_bytes_total: Arc::new(AtomicUsize::new(0)),
-            control_churn_bytes_total: Arc::new(AtomicUsize::new(0)),
-            #[cfg(windows)]
-            terminal_input_relay_stop: Arc::new(AtomicBool::new(false)),
-            #[cfg(windows)]
-            terminal_input_relay_active: Arc::new(AtomicBool::new(false)),
-            #[cfg(windows)]
-            terminal_input_relay_worker: Mutex::new(None),
-        })
+        let inner = CoreNativePtyProcess::new(argv, cwd, env_pairs, rows, cols, nice)
+            .map_err(Self::pty_err_to_py)?;
+        Ok(Self { inner })
     }
 
     #[inline(never)]
     fn start(&self) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_start_public(self)
+        self.inner.start_impl().map_err(Self::pty_err_to_py)
     }
 
     #[pyo3(signature = (timeout=None))]
     fn read_chunk(&self, py: Python<'_>, timeout: Option<f64>) -> PyResult<Py<PyAny>> {
-        // Wait for a chunk WITHOUT holding the GIL. The previous version
-        // called `condvar.wait()` while still holding the GIL, which starved
-        // every other Python thread for the duration of the wait. With a
-        // 100ms read poll loop, that meant the main thread could only run
-        // for a few microseconds every 100ms — turning ordinary calls like
-        // `os.path.realpath` into ~430ms operations and producing apparent
-        // deadlocks during pytest failure formatting.
-        enum WaitOutcome {
-            Chunk(Vec<u8>),
-            Closed,
-            Timeout,
-        }
-
-        let outcome = py.allow_threads(|| -> WaitOutcome {
-            let deadline = timeout.map(|secs| Instant::now() + Duration::from_secs_f64(secs));
-            let mut guard = self.reader.state.lock().expect("pty read mutex poisoned");
-            loop {
-                if let Some(chunk) = guard.chunks.pop_front() {
-                    return WaitOutcome::Chunk(chunk);
-                }
-                if guard.closed {
-                    return WaitOutcome::Closed;
-                }
-                match deadline {
-                    Some(deadline) => {
-                        let now = Instant::now();
-                        if now >= deadline {
-                            return WaitOutcome::Timeout;
-                        }
-                        let wait = deadline.saturating_duration_since(now);
-                        let result = self
-                            .reader
-                            .condvar
-                            .wait_timeout(guard, wait)
-                            .expect("pty read mutex poisoned");
-                        guard = result.0;
-                    }
-                    None => {
-                        guard = self
-                            .reader
-                            .condvar
-                            .wait(guard)
-                            .expect("pty read mutex poisoned");
-                    }
-                }
-            }
-        });
-
-        match outcome {
-            WaitOutcome::Chunk(chunk) => Ok(PyBytes::new(py, &chunk).into_any().unbind()),
-            WaitOutcome::Closed => Err(PyRuntimeError::new_err("Pseudo-terminal stream is closed")),
-            WaitOutcome::Timeout => Err(PyTimeoutError::new_err(
+        let result = py.allow_threads(|| self.inner.read_chunk_impl(timeout));
+        match result {
+            Ok(Some(chunk)) => Ok(PyBytes::new(py, &chunk).into_any().unbind()),
+            Ok(None) => Err(PyTimeoutError::new_err(
                 "No pseudo-terminal output available before timeout",
             )),
+            Err(e) => Err(Self::pty_err_to_py(e)),
         }
     }
 
     #[pyo3(signature = (timeout=None))]
     fn wait_for_reader_closed(&self, py: Python<'_>, timeout: Option<f64>) -> PyResult<bool> {
-        let closed = py.allow_threads(|| {
-            let deadline = timeout.map(|secs| Instant::now() + Duration::from_secs_f64(secs));
-            let mut guard = self.reader.state.lock().expect("pty read mutex poisoned");
-            loop {
-                if guard.closed {
-                    return true;
-                }
-                match deadline {
-                    Some(deadline) => {
-                        let now = Instant::now();
-                        if now >= deadline {
-                            return false;
-                        }
-                        let wait = deadline.saturating_duration_since(now);
-                        let result = self
-                            .reader
-                            .condvar
-                            .wait_timeout(guard, wait)
-                            .expect("pty read mutex poisoned");
-                        guard = result.0;
-                    }
-                    None => {
-                        guard = self
-                            .reader
-                            .condvar
-                            .wait(guard)
-                            .expect("pty read mutex poisoned");
-                    }
-                }
-            }
-        });
-        Ok(closed)
+        Ok(py.allow_threads(|| self.inner.wait_for_reader_closed_impl(timeout)))
     }
 
     #[pyo3(signature = (data, submit=false))]
     fn write(&self, data: &[u8], submit: bool) -> PyResult<()> {
-        self.write_impl(data, submit)
+        self.inner.write_impl(data, submit).map_err(Self::pty_err_to_py)
     }
 
     fn respond_to_queries(&self, data: &[u8]) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_respond_to_queries_public(self, data)
+        self.inner.respond_to_queries_impl(data).map_err(Self::pty_err_to_py)
     }
 
     #[inline(never)]
     fn resize(&self, rows: u16, cols: u16) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_resize_public(self, rows, cols)
+        self.inner.resize_impl(rows, cols).map_err(Self::pty_err_to_py)
     }
 
     #[inline(never)]
     fn send_interrupt(&self) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_send_interrupt_public(self)
+        self.inner.send_interrupt_impl().map_err(Self::pty_err_to_py)
     }
 
     fn poll(&self) -> PyResult<Option<i32>> {
-        poll_pty_process(&self.handles, &self.returncode).map_err(to_py_err)
+        core_pty::poll_pty_process(&self.inner.handles, &self.inner.returncode)
+            .map_err(to_py_err)
     }
 
     #[pyo3(signature = (timeout=None))]
     #[inline(never)]
     fn wait(&self, timeout: Option<f64>) -> PyResult<i32> {
-        public_symbols::rp_native_pty_process_wait_public(self, timeout)
+        self.inner.wait_impl(timeout).map_err(Self::pty_err_to_py)
     }
 
     #[inline(never)]
     fn terminate(&self) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_terminate_public(self)
+        self.inner.terminate_impl().map_err(Self::pty_err_to_py)
     }
 
     #[inline(never)]
     fn kill(&self) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_kill_public(self)
+        self.inner.kill_impl().map_err(Self::pty_err_to_py)
     }
 
     #[inline(never)]
     fn terminate_tree(&self) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_terminate_tree_public(self)
+        self.inner.terminate_tree_impl().map_err(Self::pty_err_to_py)
     }
 
     #[inline(never)]
     fn kill_tree(&self) -> PyResult<()> {
-        public_symbols::rp_native_pty_process_kill_tree_public(self)
+        self.inner.kill_tree_impl().map_err(Self::pty_err_to_py)
     }
 
     fn start_terminal_input_relay(&self) -> PyResult<()> {
-        self.start_terminal_input_relay_impl()
+        self.start_terminal_input_relay_py()
     }
 
     fn stop_terminal_input_relay(&self) {
-        self.stop_terminal_input_relay_impl();
+        self.inner.stop_terminal_input_relay_impl();
     }
 
     fn terminal_input_relay_active(&self) -> bool {
         #[cfg(windows)]
         {
-            self.terminal_input_relay_active.load(Ordering::Acquire)
+            self.inner.terminal_input_relay_active.load(Ordering::Acquire)
         }
 
         #[cfg(not(windows))]
@@ -2918,29 +1727,25 @@ impl NativePtyProcess {
     }
 
     fn pty_input_bytes_total(&self) -> usize {
-        self.input_bytes_total.load(Ordering::Acquire)
+        self.inner.pty_input_bytes_total()
     }
 
     fn pty_newline_events_total(&self) -> usize {
-        self.newline_events_total.load(Ordering::Acquire)
+        self.inner.pty_newline_events_total()
     }
 
     fn pty_submit_events_total(&self) -> usize {
-        self.submit_events_total.load(Ordering::Acquire)
+        self.inner.pty_submit_events_total()
     }
 
-    /// Visible (non-control) output bytes tracked by the reader thread.
     fn pty_output_bytes_total(&self) -> usize {
-        self.output_bytes_total.load(Ordering::Acquire)
+        self.inner.pty_output_bytes_total()
     }
 
-    /// Control churn bytes (ANSI escapes, BS, CR, DEL) tracked by the reader thread.
     fn pty_control_churn_bytes_total(&self) -> usize {
-        self.control_churn_bytes_total.load(Ordering::Acquire)
+        self.inner.pty_control_churn_bytes_total()
     }
 
-    /// Wait for exit then drain remaining output.  Entire operation runs
-    /// in Rust with the GIL released.
     #[pyo3(signature = (timeout=None, drain_timeout=2.0))]
     fn wait_and_drain(
         &self,
@@ -2949,56 +1754,28 @@ impl NativePtyProcess {
         drain_timeout: f64,
     ) -> PyResult<i32> {
         py.allow_threads(|| {
-            // Wait for exit.
-            let code = self.wait_impl(timeout)?;
-            // Drain: wait for reader thread to close.
-            let deadline = Instant::now() + Duration::from_secs_f64(drain_timeout.max(0.0));
-            let mut guard = self.reader.state.lock().expect("pty read mutex poisoned");
-            while !guard.closed {
-                let remaining = deadline.saturating_duration_since(Instant::now());
-                if remaining.is_zero() {
-                    break;
-                }
-                let result = self
-                    .reader
-                    .condvar
-                    .wait_timeout(guard, remaining)
-                    .expect("pty read mutex poisoned");
-                guard = result.0;
-            }
-            Ok(code)
+            self.inner
+                .wait_and_drain_impl(timeout, drain_timeout)
+                .map_err(Self::pty_err_to_py)
         })
     }
 
-    /// Enable/disable echoing PTY output to stdout from the reader thread.
     fn set_echo(&self, enabled: bool) {
-        self.echo.store(enabled, Ordering::Release);
+        self.inner.set_echo(enabled);
     }
 
     fn echo_enabled(&self) -> bool {
-        self.echo.load(Ordering::Acquire)
+        self.inner.echo_enabled()
     }
 
-    /// Attach an idle detector so the reader thread feeds it directly.
     fn attach_idle_detector(&self, detector: &NativeIdleDetector) {
-        let mut guard = self
-            .idle_detector
-            .lock()
-            .expect("idle detector mutex poisoned");
-        *guard = Some(Arc::clone(&detector.core));
+        self.inner.attach_idle_detector(&detector.core);
     }
 
-    /// Detach the idle detector from the reader thread.
     fn detach_idle_detector(&self) {
-        let mut guard = self
-            .idle_detector
-            .lock()
-            .expect("idle detector mutex poisoned");
-        *guard = None;
+        self.inner.detach_idle_detector();
     }
 
-    /// Wait for idle entirely in Rust.  The reader thread feeds the
-    /// detector directly — no Python pumping needed.
     #[pyo3(signature = (detector, timeout=None))]
     fn wait_for_idle(
         &self,
@@ -3007,22 +1784,15 @@ impl NativePtyProcess {
         timeout: Option<f64>,
     ) -> PyResult<(bool, String, f64, Option<i32>)> {
         // Wire the detector into the reader thread.
-        {
-            let mut guard = self
-                .idle_detector
-                .lock()
-                .expect("idle detector mutex poisoned");
-            *guard = Some(Arc::clone(&detector.core));
-        }
+        self.inner.attach_idle_detector(&detector.core);
 
         // Spawn exit watcher that marks the detector on process exit.
-        let handles = Arc::clone(&self.handles);
-        let returncode = Arc::clone(&self.returncode);
+        let handles = Arc::clone(&self.inner.handles);
+        let returncode = Arc::clone(&self.inner.returncode);
         let core = Arc::clone(&detector.core);
         let exit_watcher = thread::spawn(move || loop {
-            match poll_pty_process(&handles, &returncode) {
+            match core_pty::poll_pty_process(&handles, &returncode) {
                 Ok(Some(code)) => {
-                    // Heuristic: codes typically used for keyboard interrupt
                     let interrupted = code == -2 || code == 130;
                     core.mark_exit(code, interrupted);
                     return;
@@ -3033,51 +1803,20 @@ impl NativePtyProcess {
             thread::sleep(Duration::from_millis(1));
         });
 
-        // Block in Rust (GIL released) until idle, exit, or timeout.
         let result = py.allow_threads(|| detector.core.wait(timeout));
 
-        // Detach detector from reader thread.
-        {
-            let mut guard = self
-                .idle_detector
-                .lock()
-                .expect("idle detector mutex poisoned");
-            *guard = None;
-        }
+        self.inner.detach_idle_detector();
         let _ = exit_watcher.join();
         Ok(result)
     }
 
     #[getter]
     fn pid(&self) -> PyResult<Option<u32>> {
-        let guard = self.handles.lock().expect("pty handles mutex poisoned");
-        if let Some(handles) = guard.as_ref() {
-            #[cfg(unix)]
-            if let Some(pid) = handles.master.process_group_leader() {
-                if let Ok(pid) = u32::try_from(pid) {
-                    return Ok(Some(pid));
-                }
-            }
-            return Ok(handles.child.process_id());
-        }
-        Ok(None)
+        self.inner.pid().map_err(Self::pty_err_to_py)
     }
 
     fn close(&self, py: Python<'_>) -> PyResult<()> {
-        // Release the GIL while waiting on the child — otherwise the wait
-        // blocks every other Python thread (including the one that may need
-        // to drop additional references for the child to actually exit).
-        public_symbols::rp_native_pty_process_close_public(self, py)
-    }
-}
-
-impl Drop for NativePtyProcess {
-    fn drop(&mut self) {
-        // Drop runs under the GIL during PyO3 finalization. Calling
-        // `close_impl` here would block the interpreter on `child.wait()`
-        // and deadlock with the background reader thread. Use the
-        // non-blocking teardown instead.
-        public_symbols::rp_native_pty_process_close_nonblocking_public(self);
+        py.allow_threads(|| self.inner.close_impl().map_err(Self::pty_err_to_py))
     }
 }
 
@@ -3259,14 +1998,6 @@ impl NativePtyBuffer {
 }
 
 impl NativeTerminalInput {
-    fn next_event(&self) -> Option<TerminalInputEventRecord> {
-        self.state
-            .lock()
-            .expect("terminal input mutex poisoned")
-            .events
-            .pop_front()
-    }
-
     fn event_to_py(
         py: Python<'_>,
         event: TerminalInputEventRecord,
@@ -3290,164 +2021,19 @@ impl NativeTerminalInput {
         py: Python<'_>,
         timeout: Option<f64>,
     ) -> PyResult<TerminalInputEventRecord> {
-        enum WaitOutcome {
-            Event(TerminalInputEventRecord),
-            Closed,
-            Timeout,
-        }
-
-        let state = Arc::clone(&self.state);
-        let condvar = Arc::clone(&self.condvar);
-        let outcome = py.allow_threads(move || -> WaitOutcome {
-            let deadline = timeout.map(|secs| Instant::now() + Duration::from_secs_f64(secs));
-            let mut guard = state.lock().expect("terminal input mutex poisoned");
-            loop {
-                if let Some(event) = guard.events.pop_front() {
-                    return WaitOutcome::Event(event);
-                }
-                if guard.closed {
-                    return WaitOutcome::Closed;
-                }
-                match deadline {
-                    Some(deadline) => {
-                        let now = Instant::now();
-                        if now >= deadline {
-                            return WaitOutcome::Timeout;
-                        }
-                        let wait = deadline.saturating_duration_since(now);
-                        let result = condvar
-                            .wait_timeout(guard, wait)
-                            .expect("terminal input mutex poisoned");
-                        guard = result.0;
+        py.allow_threads(|| {
+            self.inner
+                .wait_for_event(timeout)
+                .map_err(|err| match err {
+                    TerminalInputError::Closed => {
+                        PyRuntimeError::new_err("Native terminal input is closed")
                     }
-                    None => {
-                        guard = condvar.wait(guard).expect("terminal input mutex poisoned");
-                    }
-                }
-            }
-        });
-
-        match outcome {
-            WaitOutcome::Event(event) => Ok(event),
-            WaitOutcome::Closed => Err(PyRuntimeError::new_err("Native terminal input is closed")),
-            WaitOutcome::Timeout => Err(PyTimeoutError::new_err(
-                "No terminal input available before timeout",
-            )),
-        }
-    }
-
-    fn stop_impl(&self) -> PyResult<()> {
-        self.stop.store(true, Ordering::Release);
-        #[cfg(windows)]
-        append_native_terminal_input_trace_line(&format!(
-            "[{:.6}] native_terminal_input stop_requested",
-            unix_now_seconds(),
-        ));
-        if let Some(worker) = self
-            .worker
-            .lock()
-            .expect("terminal input worker mutex poisoned")
-            .take()
-        {
-            let _ = worker.join();
-        }
-        self.capturing.store(false, Ordering::Release);
-
-        #[cfg(windows)]
-        let restore_result = {
-            use winapi::um::consoleapi::SetConsoleMode;
-            use winapi::um::winnt::HANDLE;
-
-            let console = self
-                .console
-                .lock()
-                .expect("terminal input console mutex poisoned")
-                .take();
-            console.map(|capture| unsafe {
-                SetConsoleMode(capture.input_handle as HANDLE, capture.original_mode)
-            })
-        };
-
-        let mut guard = self.state.lock().expect("terminal input mutex poisoned");
-        guard.closed = true;
-        self.condvar.notify_all();
-        drop(guard);
-
-        #[cfg(windows)]
-        if let Some(result) = restore_result {
-            if result == 0 {
-                return Err(to_py_err(std::io::Error::last_os_error()));
-            }
-        }
-        Ok(())
-    }
-
-    #[cfg(windows)]
-    fn start_impl(&self) -> PyResult<()> {
-        use winapi::um::consoleapi::{GetConsoleMode, SetConsoleMode};
-        use winapi::um::handleapi::INVALID_HANDLE_VALUE;
-        use winapi::um::processenv::GetStdHandle;
-        use winapi::um::winbase::STD_INPUT_HANDLE;
-
-        let mut worker_guard = self
-            .worker
-            .lock()
-            .expect("terminal input worker mutex poisoned");
-        if worker_guard.is_some() {
-            return Ok(());
-        }
-
-        let input_handle = unsafe { GetStdHandle(STD_INPUT_HANDLE) };
-        if input_handle.is_null() || input_handle == INVALID_HANDLE_VALUE {
-            return Err(to_py_err(std::io::Error::last_os_error()));
-        }
-
-        let mut original_mode = 0u32;
-        let got_mode = unsafe { GetConsoleMode(input_handle, &mut original_mode) };
-        if got_mode == 0 {
-            return Err(PyRuntimeError::new_err(
-                "NativeTerminalInput requires an attached Windows console stdin",
-            ));
-        }
-
-        let active_mode = native_terminal_input_mode(original_mode);
-        let set_mode = unsafe { SetConsoleMode(input_handle, active_mode) };
-        if set_mode == 0 {
-            return Err(to_py_err(std::io::Error::last_os_error()));
-        }
-        append_native_terminal_input_trace_line(&format!(
-            "[{:.6}] native_terminal_input start handle={} original_mode={:#010x} active_mode={:#010x}",
-            unix_now_seconds(),
-            input_handle as usize,
-            original_mode,
-            active_mode,
-        ));
-
-        self.stop.store(false, Ordering::Release);
-        self.capturing.store(true, Ordering::Release);
-        {
-            let mut state = self.state.lock().expect("terminal input mutex poisoned");
-            state.events.clear();
-            state.closed = false;
-        }
-        *self
-            .console
-            .lock()
-            .expect("terminal input console mutex poisoned") = Some(ActiveTerminalInputCapture {
-            input_handle: input_handle as usize,
-            original_mode,
-            active_mode,
-        });
-
-        let state = Arc::clone(&self.state);
-        let condvar = Arc::clone(&self.condvar);
-        let stop = Arc::clone(&self.stop);
-        let capturing = Arc::clone(&self.capturing);
-        let input_handle_raw = input_handle as usize;
-        *worker_guard = Some(thread::spawn(move || {
-            native_terminal_input_worker(input_handle_raw, state, condvar, stop, capturing);
-        }));
-        Ok(())
+                    TerminalInputError::Timeout => PyTimeoutError::new_err(
+                        "No terminal input available before timeout",
+                    ),
+                    other => to_py_err(other),
+                })
+        })
     }
 }
 
@@ -3507,23 +2093,14 @@ impl NativeTerminalInput {
     #[new]
     fn new() -> Self {
         Self {
-            state: Arc::new(Mutex::new(TerminalInputState {
-                events: VecDeque::new(),
-                closed: true,
-            })),
-            condvar: Arc::new(Condvar::new()),
-            stop: Arc::new(AtomicBool::new(false)),
-            capturing: Arc::new(AtomicBool::new(false)),
-            worker: Mutex::new(None),
-            #[cfg(windows)]
-            console: Mutex::new(None),
+            inner: TerminalInputCore::new(),
         }
     }
 
     fn start(&self) -> PyResult<()> {
         #[cfg(windows)]
         {
-            self.start_impl()
+            self.inner.start_impl().map_err(to_py_err)
         }
 
         #[cfg(not(windows))]
@@ -3535,61 +2112,30 @@ impl NativeTerminalInput {
     }
 
     fn stop(&self, py: Python<'_>) -> PyResult<()> {
-        py.allow_threads(|| self.stop_impl())
+        py.allow_threads(|| self.inner.stop_impl().map_err(to_py_err))
     }
 
     fn close(&self, py: Python<'_>) -> PyResult<()> {
-        py.allow_threads(|| self.stop_impl())
+        py.allow_threads(|| self.inner.stop_impl().map_err(to_py_err))
     }
 
     fn available(&self) -> bool {
-        !self
-            .state
-            .lock()
-            .expect("terminal input mutex poisoned")
-            .events
-            .is_empty()
+        self.inner.available()
     }
 
     #[getter]
     fn capturing(&self) -> bool {
-        self.capturing.load(Ordering::Acquire)
+        self.inner.capturing()
     }
 
     #[getter]
     fn original_console_mode(&self) -> Option<u32> {
-        #[cfg(windows)]
-        {
-            return self
-                .console
-                .lock()
-                .expect("terminal input console mutex poisoned")
-                .as_ref()
-                .map(|capture| capture.original_mode);
-        }
-
-        #[cfg(not(windows))]
-        {
-            None
-        }
+        self.inner.original_console_mode()
     }
 
     #[getter]
     fn active_console_mode(&self) -> Option<u32> {
-        #[cfg(windows)]
-        {
-            return self
-                .console
-                .lock()
-                .expect("terminal input console mutex poisoned")
-                .as_ref()
-                .map(|capture| capture.active_mode);
-        }
-
-        #[cfg(not(windows))]
-        {
-            None
-        }
+        self.inner.active_console_mode()
     }
 
     #[pyo3(signature = (timeout=None))]
@@ -3606,10 +2152,11 @@ impl NativeTerminalInput {
         &self,
         py: Python<'_>,
     ) -> PyResult<Option<Py<NativeTerminalInputEvent>>> {
-        if let Some(event) = self.next_event() {
+        if let Some(event) = self.inner.next_event() {
             return Self::event_to_py(py, event).map(Some);
         }
         if self
+            .inner
             .state
             .lock()
             .expect("terminal input mutex poisoned")
@@ -3627,10 +2174,11 @@ impl NativeTerminalInput {
     }
 
     fn read_non_blocking(&self, py: Python<'_>) -> PyResult<Option<Py<PyAny>>> {
-        if let Some(event) = self.next_event() {
+        if let Some(event) = self.inner.next_event() {
             return Ok(Some(PyBytes::new(py, &event.data).into_any().unbind()));
         }
         if self
+            .inner
             .state
             .lock()
             .expect("terminal input mutex poisoned")
@@ -3642,19 +2190,17 @@ impl NativeTerminalInput {
     }
 
     fn drain(&self, py: Python<'_>) -> Vec<Py<PyAny>> {
-        let mut guard = self.state.lock().expect("terminal input mutex poisoned");
-        guard
-            .events
-            .drain(..)
+        self.inner
+            .drain_events()
+            .into_iter()
             .map(|event| PyBytes::new(py, &event.data).into_any().unbind())
             .collect()
     }
 
     fn drain_events(&self, py: Python<'_>) -> PyResult<Vec<Py<NativeTerminalInputEvent>>> {
-        let mut guard = self.state.lock().expect("terminal input mutex poisoned");
-        guard
-            .events
-            .drain(..)
+        self.inner
+            .drain_events()
+            .into_iter()
             .map(|event| Self::event_to_py(py, event))
             .collect()
     }
@@ -3673,10 +2219,8 @@ impl NativeTerminalInput {
         // Block (releasing the GIL) until the first event arrives.
         let first = self.wait_for_event(py, timeout)?;
 
-        // Drain everything else already queued — single lock acquisition.
-        let mut guard = self.state.lock().expect("terminal input mutex poisoned");
-        let remaining: Vec<TerminalInputEventRecord> = guard.events.drain(..).collect();
-        drop(guard);
+        // Drain everything else already queued.
+        let remaining = self.inner.drain_events();
 
         // Merge all data into one buffer.
         let capacity = first.data.len() + remaining.iter().map(|e| e.data.len()).sum::<usize>();
@@ -3692,11 +2236,7 @@ impl NativeTerminalInput {
     }
 }
 
-impl Drop for NativeTerminalInput {
-    fn drop(&mut self) {
-        let _ = self.stop_impl();
-    }
-}
+// Drop is now handled by TerminalInputCore's Drop impl
 
 impl NativeRunningProcess {
     fn start_impl(&self) -> PyResult<()> {
@@ -3964,216 +2504,31 @@ impl NativeIdleDetector {
     }
 }
 
-fn control_churn_bytes(data: &[u8]) -> usize {
-    let mut total = 0;
-    let mut index = 0;
-    while index < data.len() {
-        let byte = data[index];
-        if byte == 0x1B {
-            let start = index;
-            index += 1;
-            if index < data.len() && data[index] == b'[' {
-                index += 1;
-                while index < data.len() {
-                    let current = data[index];
-                    index += 1;
-                    if (0x40..=0x7E).contains(&current) {
-                        break;
-                    }
-                }
-            }
-            total += index - start;
-            continue;
-        }
-        if matches!(byte, 0x08 | 0x0D | 0x7F) {
-            total += 1;
-        }
-        index += 1;
-    }
-    total
-}
-
-fn command_builder_from_argv(argv: &[String]) -> CommandBuilder {
-    let mut command = CommandBuilder::new(&argv[0]);
-    if argv.len() > 1 {
-        command.args(
-            argv[1..]
-                .iter()
-                .map(OsString::from)
-                .collect::<Vec<OsString>>(),
-        );
-    }
-    command
-}
-
-#[inline(never)]
-fn spawn_pty_reader(
-    mut reader: Box<dyn Read + Send>,
-    shared: Arc<PtyReadShared>,
-    echo: Arc<AtomicBool>,
-    idle_detector: Arc<Mutex<Option<Arc<IdleDetectorCore>>>>,
-    output_bytes_total: Arc<AtomicUsize>,
-    control_churn_bytes_total: Arc<AtomicUsize>,
-) {
-    running_process_core::rp_rust_debug_scope!("running_process_py::spawn_pty_reader");
-    let idle_detector_snapshot = idle_detector
-        .lock()
-        .expect("idle detector mutex poisoned")
-        .clone();
-    let mut chunk = vec![0_u8; 65536];
-    loop {
-        match reader.read(&mut chunk) {
-            Ok(0) => break,
-            Ok(n) => {
-                let data = &chunk[..n];
-
-                // Output accounting (no GIL needed).
-                let churn = control_churn_bytes(data);
-                let visible = data.len().saturating_sub(churn);
-                output_bytes_total.fetch_add(visible, Ordering::Relaxed);
-                control_churn_bytes_total.fetch_add(churn, Ordering::Relaxed);
-
-                // Echo to stdout if enabled (no GIL needed).
-                if echo.load(Ordering::Relaxed) {
-                    let _ = std::io::stdout().write_all(data);
-                    let _ = std::io::stdout().flush();
-                }
-
-                // Feed idle detector directly (no GIL needed).
-                if let Some(ref detector) = idle_detector_snapshot {
-                    detector.record_output(data);
-                }
-
-                let mut guard = shared.state.lock().expect("pty read mutex poisoned");
-                guard.chunks.push_back(data.to_vec());
-                shared.condvar.notify_all();
-            }
-            Err(err) if err.kind() == std::io::ErrorKind::Interrupted => continue,
-            Err(err) if err.kind() == std::io::ErrorKind::WouldBlock => {
-                thread::sleep(Duration::from_millis(10));
-                continue;
-            }
-            Err(_) => break,
-        }
-    }
-    let mut guard = shared.state.lock().expect("pty read mutex poisoned");
-    guard.closed = true;
-    shared.condvar.notify_all();
-}
-
-fn portable_exit_code(status: portable_pty::ExitStatus) -> i32 {
-    if let Some(signal) = status.signal() {
-        let signal = signal.to_ascii_lowercase();
-        if signal.contains("interrupt") {
-            return -2;
-        }
-        if signal.contains("terminated") {
-            return -15;
-        }
-        if signal.contains("killed") {
-            return -9;
-        }
-    }
-    status.exit_code() as i32
-}
-
-#[cfg(windows)]
-#[inline(never)]
-fn assign_child_to_windows_kill_on_close_job(
-    handle: Option<std::os::windows::io::RawHandle>,
-) -> PyResult<WindowsJobHandle> {
-    running_process_core::rp_rust_debug_scope!(
-        "running_process_py::assign_child_to_windows_kill_on_close_job"
-    );
-    use std::mem::zeroed;
-
-    use winapi::shared::minwindef::FALSE;
-    use winapi::um::handleapi::INVALID_HANDLE_VALUE;
-    use winapi::um::jobapi2::{
-        AssignProcessToJobObject, CreateJobObjectW, SetInformationJobObject,
-    };
-    use winapi::um::winnt::{
-        JobObjectExtendedLimitInformation, JOBOBJECT_EXTENDED_LIMIT_INFORMATION,
-        JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE,
-    };
-
-    let Some(handle) = handle else {
-        return Err(PyRuntimeError::new_err(
-            "Pseudo-terminal child does not expose a Windows process handle",
-        ));
-    };
-
-    let job = unsafe { CreateJobObjectW(std::ptr::null_mut(), std::ptr::null()) };
-    if job.is_null() || job == INVALID_HANDLE_VALUE {
-        return Err(to_py_err(std::io::Error::last_os_error()));
-    }
-
-    let mut info: JOBOBJECT_EXTENDED_LIMIT_INFORMATION = unsafe { zeroed() };
-    info.BasicLimitInformation.LimitFlags = JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE;
-    let result = unsafe {
-        SetInformationJobObject(
-            job,
-            JobObjectExtendedLimitInformation,
-            (&mut info as *mut JOBOBJECT_EXTENDED_LIMIT_INFORMATION).cast(),
-            std::mem::size_of::<JOBOBJECT_EXTENDED_LIMIT_INFORMATION>() as u32,
-        )
-    };
-    if result == FALSE {
-        let err = std::io::Error::last_os_error();
-        unsafe {
-            winapi::um::handleapi::CloseHandle(job);
-        }
-        return Err(to_py_err(err));
-    }
-
-    let result = unsafe { AssignProcessToJobObject(job, handle.cast()) };
-    if result == FALSE {
-        let err = std::io::Error::last_os_error();
-        unsafe {
-            winapi::um::handleapi::CloseHandle(job);
-        }
-        return Err(to_py_err(err));
-    }
-
-    Ok(WindowsJobHandle(job as usize))
-}
-
-#[cfg(windows)]
-#[inline(never)]
-fn apply_windows_pty_priority(
-    handle: Option<std::os::windows::io::RawHandle>,
-    nice: Option<i32>,
-) -> PyResult<()> {
-    running_process_core::rp_rust_debug_scope!("running_process_py::apply_windows_pty_priority");
-    use winapi::um::processthreadsapi::SetPriorityClass;
-    use winapi::um::winbase::{
-        ABOVE_NORMAL_PRIORITY_CLASS, BELOW_NORMAL_PRIORITY_CLASS, HIGH_PRIORITY_CLASS,
-        IDLE_PRIORITY_CLASS,
-    };
-
-    let Some(handle) = handle else {
-        return Ok(());
-    };
-    let flags = match nice {
-        Some(value) if value >= 15 => IDLE_PRIORITY_CLASS,
-        Some(value) if value >= 1 => BELOW_NORMAL_PRIORITY_CLASS,
-        Some(value) if value <= -15 => HIGH_PRIORITY_CLASS,
-        Some(value) if value <= -1 => ABOVE_NORMAL_PRIORITY_CLASS,
-        _ => 0,
-    };
-    if flags == 0 {
-        return Ok(());
-    }
-    let result = unsafe { SetPriorityClass(handle.cast(), flags) };
-    if result == 0 {
-        return Err(to_py_err(std::io::Error::last_os_error()));
-    }
-    Ok(())
-}
+// PTY helper functions (control_churn_bytes, command_builder_from_argv,
+// spawn_pty_reader, portable_exit_code, assign_child_to_windows_kill_on_close_job,
+// apply_windows_pty_priority) are now in running_process_core::pty
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use running_process_core::pty::{
+        self as core_pty, NativePtyHandles, NativePtyProcess as CoreNativePtyProcess, PtyReadShared,
+        PtyReadState,
+    };
+    #[cfg(windows)]
+    use running_process_core::pty::{
+        assign_child_to_windows_kill_on_close_job, apply_windows_pty_priority, WindowsJobHandle,
+    };
+    #[cfg(windows)]
+    use running_process_core::pty::terminal_input::{
+        TerminalInputState,
+        native_terminal_input_trace_target,
+        format_terminal_input_bytes,
+        native_terminal_input_mode, terminal_input_modifier_parameter,
+        repeat_terminal_input_bytes, repeated_modified_sequence, repeated_tilde_sequence,
+        control_character_for_unicode,
+        translate_console_key_event,
+    };
 
     #[cfg(windows)]
     use winapi::um::wincon::{
@@ -4300,57 +2655,57 @@ mod tests {
 
     #[test]
     fn control_churn_bytes_empty() {
-        assert_eq!(control_churn_bytes(b""), 0);
+        assert_eq!(core_pty::control_churn_bytes(b""), 0);
     }
 
     #[test]
     fn control_churn_bytes_plain_text() {
-        assert_eq!(control_churn_bytes(b"hello world"), 0);
+        assert_eq!(core_pty::control_churn_bytes(b"hello world"), 0);
     }
 
     #[test]
     fn control_churn_bytes_ansi_csi_sequence() {
         // \x1b[31m = 5 bytes of control churn, \x1b[0m = 4 bytes
-        assert_eq!(control_churn_bytes(b"\x1b[31mhello\x1b[0m"), 9);
+        assert_eq!(core_pty::control_churn_bytes(b"\x1b[31mhello\x1b[0m"), 9);
     }
 
     #[test]
     fn control_churn_bytes_backspace_cr_del() {
-        assert_eq!(control_churn_bytes(b"\x08\x0D\x7F"), 3);
+        assert_eq!(core_pty::control_churn_bytes(b"\x08\x0D\x7F"), 3);
     }
 
     #[test]
     fn control_churn_bytes_bare_escape() {
         // Bare ESC with no CSI sequence following
-        assert_eq!(control_churn_bytes(b"\x1b"), 1);
+        assert_eq!(core_pty::control_churn_bytes(b"\x1b"), 1);
     }
 
     #[test]
     fn control_churn_bytes_mixed() {
         // \x1b[J = 3 bytes CSI + 1 byte BS = 4
-        assert_eq!(control_churn_bytes(b"ok\x1b[Jmore\x08"), 4);
+        assert_eq!(core_pty::control_churn_bytes(b"ok\x1b[Jmore\x08"), 4);
     }
 
     // ── input_contains_newline tests ──
 
     #[test]
     fn input_contains_newline_cr() {
-        assert!(input_contains_newline(b"hello\rworld"));
+        assert!(core_pty::input_contains_newline(b"hello\rworld"));
     }
 
     #[test]
     fn input_contains_newline_lf() {
-        assert!(input_contains_newline(b"hello\nworld"));
+        assert!(core_pty::input_contains_newline(b"hello\nworld"));
     }
 
     #[test]
     fn input_contains_newline_none() {
-        assert!(!input_contains_newline(b"hello world"));
+        assert!(!core_pty::input_contains_newline(b"hello world"));
     }
 
     #[test]
     fn input_contains_newline_empty() {
-        assert!(!input_contains_newline(b""));
+        assert!(!core_pty::input_contains_newline(b""));
     }
 
     // ── is_ignorable_process_control_error tests ──
@@ -4385,28 +2740,28 @@ mod tests {
     #[test]
     #[cfg(windows)]
     fn windows_terminal_input_payload_passthrough() {
-        let result = windows_terminal_input_payload(b"hello");
+        let result = core_pty::windows_terminal_input_payload(b"hello");
         assert_eq!(result, b"hello");
     }
 
     #[test]
     #[cfg(windows)]
     fn windows_terminal_input_payload_lone_lf_becomes_cr() {
-        let result = windows_terminal_input_payload(b"\n");
+        let result = core_pty::windows_terminal_input_payload(b"\n");
         assert_eq!(result, b"\r");
     }
 
     #[test]
     #[cfg(windows)]
     fn windows_terminal_input_payload_crlf_preserved() {
-        let result = windows_terminal_input_payload(b"\r\n");
+        let result = core_pty::windows_terminal_input_payload(b"\r\n");
         assert_eq!(result, b"\r\n");
     }
 
     #[test]
     #[cfg(windows)]
     fn windows_terminal_input_payload_lone_cr_preserved() {
-        let result = windows_terminal_input_payload(b"\r");
+        let result = core_pty::windows_terminal_input_payload(b"\r");
         assert_eq!(result, b"\r");
     }
 
@@ -5133,14 +3488,14 @@ mod tests {
 
     #[test]
     fn portable_exit_code_normal_exit_zero() {
-        let status = portable_pty::ExitStatus::with_exit_code(0);
-        assert_eq!(portable_exit_code(status), 0);
+        let status = running_process_core::pty::reexports::portable_pty::ExitStatus::with_exit_code(0);
+        assert_eq!(core_pty::portable_exit_code(status), 0);
     }
 
     #[test]
     fn portable_exit_code_normal_exit_nonzero() {
-        let status = portable_pty::ExitStatus::with_exit_code(42);
-        assert_eq!(portable_exit_code(status), 42);
+        let status = running_process_core::pty::reexports::portable_pty::ExitStatus::with_exit_code(42);
+        assert_eq!(core_pty::portable_exit_code(status), 42);
     }
 
     // ── record_pty_input_metrics tests ──
@@ -5151,7 +3506,7 @@ mod tests {
         let newline_events = Arc::new(AtomicUsize::new(0));
         let submit_events = Arc::new(AtomicUsize::new(0));
 
-        record_pty_input_metrics(
+        core_pty::record_pty_input_metrics(
             &input_bytes,
             &newline_events,
             &submit_events,
@@ -5170,7 +3525,7 @@ mod tests {
         let newline_events = Arc::new(AtomicUsize::new(0));
         let submit_events = Arc::new(AtomicUsize::new(0));
 
-        record_pty_input_metrics(
+        core_pty::record_pty_input_metrics(
             &input_bytes,
             &newline_events,
             &submit_events,
@@ -5189,7 +3544,7 @@ mod tests {
         let newline_events = Arc::new(AtomicUsize::new(0));
         let submit_events = Arc::new(AtomicUsize::new(0));
 
-        record_pty_input_metrics(&input_bytes, &newline_events, &submit_events, b"\r", true);
+        core_pty::record_pty_input_metrics(&input_bytes, &newline_events, &submit_events, b"\r", true);
 
         assert_eq!(input_bytes.load(Ordering::Acquire), 1);
         assert_eq!(newline_events.load(Ordering::Acquire), 1);
@@ -5202,8 +3557,8 @@ mod tests {
         let newline_events = Arc::new(AtomicUsize::new(0));
         let submit_events = Arc::new(AtomicUsize::new(0));
 
-        record_pty_input_metrics(&input_bytes, &newline_events, &submit_events, b"ab", false);
-        record_pty_input_metrics(&input_bytes, &newline_events, &submit_events, b"cd\n", true);
+        core_pty::record_pty_input_metrics(&input_bytes, &newline_events, &submit_events, b"ab", false);
+        core_pty::record_pty_input_metrics(&input_bytes, &newline_events, &submit_events, b"cd\n", true);
 
         assert_eq!(input_bytes.load(Ordering::Acquire), 5);
         assert_eq!(newline_events.load(Ordering::Acquire), 1);
@@ -5215,14 +3570,14 @@ mod tests {
     #[test]
     fn store_pty_returncode_sets_value() {
         let returncode = Arc::new(Mutex::new(None));
-        store_pty_returncode(&returncode, 42);
+        core_pty::store_pty_returncode(&returncode, 42);
         assert_eq!(*returncode.lock().unwrap(), Some(42));
     }
 
     #[test]
     fn store_pty_returncode_overwrites() {
         let returncode = Arc::new(Mutex::new(Some(1)));
-        store_pty_returncode(&returncode, 0);
+        core_pty::store_pty_returncode(&returncode, 0);
         assert_eq!(*returncode.lock().unwrap(), Some(0));
     }
 
@@ -5231,7 +3586,7 @@ mod tests {
     #[test]
     fn write_pty_input_not_connected() {
         let handles: Arc<Mutex<Option<NativePtyHandles>>> = Arc::new(Mutex::new(None));
-        let result = write_pty_input(&handles, b"hello");
+        let result = core_pty::write_pty_input(&handles, b"hello");
         assert!(result.is_err());
         let err = result.unwrap_err();
         assert_eq!(err.kind(), std::io::ErrorKind::NotConnected);
@@ -5243,7 +3598,7 @@ mod tests {
     fn poll_pty_process_no_handles_returns_stored_code() {
         let handles: Arc<Mutex<Option<NativePtyHandles>>> = Arc::new(Mutex::new(None));
         let returncode = Arc::new(Mutex::new(Some(42)));
-        let result = poll_pty_process(&handles, &returncode).unwrap();
+        let result = core_pty::poll_pty_process(&handles, &returncode).unwrap();
         assert_eq!(result, Some(42));
     }
 
@@ -5251,7 +3606,7 @@ mod tests {
     fn poll_pty_process_no_handles_no_code() {
         let handles: Arc<Mutex<Option<NativePtyHandles>>> = Arc::new(Mutex::new(None));
         let returncode = Arc::new(Mutex::new(None));
-        let result = poll_pty_process(&handles, &returncode).unwrap();
+        let result = core_pty::poll_pty_process(&handles, &returncode).unwrap();
         assert_eq!(result, None);
     }
 
@@ -5299,14 +3654,14 @@ mod tests {
     #[test]
     fn command_builder_from_argv_single_arg() {
         let argv = vec!["echo".to_string()];
-        let _cmd = command_builder_from_argv(&argv);
+        let _cmd = core_pty::command_builder_from_argv(&argv);
         // Just ensure it doesn't panic
     }
 
     #[test]
     fn command_builder_from_argv_multi_args() {
         let argv = vec!["echo".to_string(), "hello".to_string(), "world".to_string()];
-        let _cmd = command_builder_from_argv(&argv);
+        let _cmd = core_pty::command_builder_from_argv(&argv);
         // Just ensure it doesn't panic
     }
 
@@ -5433,19 +3788,19 @@ mod tests {
     #[test]
     fn control_churn_bytes_escape_then_non_bracket() {
         // ESC followed by non-bracket character: only ESC itself is churn
-        assert_eq!(control_churn_bytes(b"\x1bO"), 1);
+        assert_eq!(core_pty::control_churn_bytes(b"\x1bO"), 1);
     }
 
     #[test]
     fn control_churn_bytes_incomplete_csi() {
         // ESC [ without terminator - counts entire remainder as churn
-        assert_eq!(control_churn_bytes(b"\x1b[123"), 5);
+        assert_eq!(core_pty::control_churn_bytes(b"\x1b[123"), 5);
     }
 
     #[test]
     fn control_churn_bytes_multiple_sequences() {
         // Two complete CSI sequences
-        assert_eq!(control_churn_bytes(b"\x1b[H\x1b[2J"), 7);
+        assert_eq!(core_pty::control_churn_bytes(b"\x1b[H\x1b[2J"), 7);
     }
 
     // ── Windows-specific additional tests ──
@@ -5456,25 +3811,25 @@ mod tests {
 
         #[test]
         fn windows_terminal_input_payload_mixed_line_endings() {
-            let result = windows_terminal_input_payload(b"a\nb\r\nc\rd");
+            let result = core_pty::windows_terminal_input_payload(b"a\nb\r\nc\rd");
             assert_eq!(result, b"a\rb\r\nc\rd");
         }
 
         #[test]
         fn windows_terminal_input_payload_consecutive_lf() {
-            let result = windows_terminal_input_payload(b"\n\n");
+            let result = core_pty::windows_terminal_input_payload(b"\n\n");
             assert_eq!(result, b"\r\r");
         }
 
         #[test]
         fn windows_terminal_input_payload_empty() {
-            let result = windows_terminal_input_payload(b"");
+            let result = core_pty::windows_terminal_input_payload(b"");
             assert!(result.is_empty());
         }
 
         #[test]
         fn windows_terminal_input_payload_no_line_endings() {
-            let result = windows_terminal_input_payload(b"hello world");
+            let result = core_pty::windows_terminal_input_payload(b"hello world");
             assert_eq!(result, b"hello world");
         }
 
@@ -5597,49 +3952,49 @@ mod tests {
 
     #[test]
     fn input_contains_newline_with_cr() {
-        assert!(input_contains_newline(b"hello\rworld"));
+        assert!(core_pty::input_contains_newline(b"hello\rworld"));
     }
 
     #[test]
     fn input_contains_newline_with_lf() {
-        assert!(input_contains_newline(b"hello\nworld"));
+        assert!(core_pty::input_contains_newline(b"hello\nworld"));
     }
 
     #[test]
     fn input_contains_newline_with_crlf() {
-        assert!(input_contains_newline(b"hello\r\nworld"));
+        assert!(core_pty::input_contains_newline(b"hello\r\nworld"));
     }
 
     #[test]
     fn input_contains_newline_without_newline() {
-        assert!(!input_contains_newline(b"hello world"));
+        assert!(!core_pty::input_contains_newline(b"hello world"));
     }
 
     // ── control_churn_bytes additional tests (iter2) ──
 
     #[test]
     fn control_churn_bytes_backspace() {
-        assert_eq!(control_churn_bytes(b"\x08"), 1);
+        assert_eq!(core_pty::control_churn_bytes(b"\x08"), 1);
     }
 
     #[test]
     fn control_churn_bytes_carriage_return() {
-        assert_eq!(control_churn_bytes(b"\x0D"), 1);
+        assert_eq!(core_pty::control_churn_bytes(b"\x0D"), 1);
     }
 
     #[test]
     fn control_churn_bytes_delete_char() {
-        assert_eq!(control_churn_bytes(b"\x7F"), 1);
+        assert_eq!(core_pty::control_churn_bytes(b"\x7F"), 1);
     }
 
     #[test]
     fn control_churn_bytes_mixed_with_text() {
-        assert_eq!(control_churn_bytes(b"hello\x0D\x1b[H"), 4);
+        assert_eq!(core_pty::control_churn_bytes(b"hello\x0D\x1b[H"), 4);
     }
 
     #[test]
     fn control_churn_bytes_plain_text_no_churn() {
-        assert_eq!(control_churn_bytes(b"hello world"), 0);
+        assert_eq!(core_pty::control_churn_bytes(b"hello world"), 0);
     }
 
     // ── system_pid tests ──
@@ -6248,7 +4603,7 @@ mod tests {
     fn terminal_input_new_starts_closed() {
         let input = NativeTerminalInput::new();
         assert!(!input.capturing());
-        let state = input.state.lock().unwrap();
+        let state = input.inner.state.lock().unwrap();
         assert!(state.closed);
         assert!(state.events.is_empty());
     }
@@ -6262,14 +4617,14 @@ mod tests {
     #[test]
     fn terminal_input_next_event_none_when_empty() {
         let input = NativeTerminalInput::new();
-        assert!(input.next_event().is_none());
+        assert!(input.inner.next_event().is_none());
     }
 
     #[test]
     fn terminal_input_inject_and_consume_event() {
         let input = NativeTerminalInput::new();
         {
-            let mut state = input.state.lock().unwrap();
+            let mut state = input.inner.state.lock().unwrap();
             state.events.push_back(TerminalInputEventRecord {
                 data: b"test".to_vec(),
                 submit: false,
@@ -6281,7 +4636,7 @@ mod tests {
             });
         }
         assert!(input.available());
-        let event = input.next_event().unwrap();
+        let event = input.inner.next_event().unwrap();
         assert_eq!(event.data, b"test");
         assert!(!input.available());
     }
@@ -6335,28 +4690,8 @@ mod tests {
 
     // ── NativePtyProcess: start_terminal_input_relay on non-windows ──
 
-    #[test]
-    #[cfg(not(windows))]
-    fn pty_process_start_terminal_input_relay_errors_on_non_windows() {
-        pyo3::prepare_freethreaded_python();
-        pyo3::Python::with_gil(|_py| {
-            let argv = vec!["echo".to_string(), "test".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
-            let result = process.start_terminal_input_relay_impl();
-            assert!(result.is_err());
-        });
-    }
-
-    #[test]
-    #[cfg(not(windows))]
-    fn pty_process_terminal_input_relay_active_false_on_non_windows() {
-        pyo3::prepare_freethreaded_python();
-        pyo3::Python::with_gil(|_py| {
-            let argv = vec!["echo".to_string(), "test".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
-            assert!(!process.terminal_input_relay_active());
-        });
-    }
+    // Terminal input relay tests are now tested through the Python wrapper
+    // since the relay logic lives in the py crate, not in core.
 
     // ── NativeProcessMetrics ──
 
@@ -6401,7 +4736,7 @@ mod tests {
     fn pty_process_empty_argv_errors() {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
-            let result = NativePtyProcess::new(vec![], None, None, 24, 80, None);
+            let result = CoreNativePtyProcess::new(vec![], None, None, 24, 80, None);
             assert!(result.is_err());
         });
     }
@@ -6418,7 +4753,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(0.1)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             let result = process.start_impl();
             assert!(result.is_err());
@@ -6486,7 +4821,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             assert!(process.pid().unwrap().is_none());
         });
     }
@@ -6501,7 +4836,7 @@ mod tests {
                 "-c".to_string(),
                 "print('hello')".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.pid().unwrap().is_some());
             let code = process.wait_impl(Some(10.0)).unwrap();
@@ -6520,9 +4855,9 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(5)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
-            assert!(process.poll().unwrap().is_none());
+            assert!(core_pty::poll_pty_process(&process.handles, &process.returncode).unwrap().is_none());
             let _ = process.close_impl();
         });
     }
@@ -6537,7 +4872,7 @@ mod tests {
                 "-c".to_string(),
                 "import sys; sys.exit(42)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             let code = process.wait_impl(Some(10.0)).unwrap();
             assert_eq!(code, 42);
@@ -6550,7 +4885,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             assert!(process.write_impl(b"test", false).is_err());
         });
     }
@@ -6565,7 +4900,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(2)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert_eq!(process.pty_input_bytes_total(), 0);
             let _ = process.write_impl(b"hello\n", false);
@@ -6587,7 +4922,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(2)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.resize_impl(40, 120).is_ok());
             let _ = process.close_impl();
@@ -6604,7 +4939,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(0.1)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.kill_impl().is_ok());
         });
@@ -6620,7 +4955,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(0.1)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.terminate_impl().is_ok());
             let _ = process.close_impl();
@@ -6633,7 +4968,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             let _ = process.wait_impl(Some(10.0));
             let _ = process.close_impl();
@@ -6651,7 +4986,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(10)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.wait_impl(Some(0.1)).is_err());
             let _ = process.close_impl();
@@ -6663,7 +4998,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             assert!(process.send_interrupt_impl().is_err());
         });
     }
@@ -6673,7 +5008,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             assert!(process.terminate_impl().is_err());
         });
     }
@@ -6683,7 +5018,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             assert!(process.kill_impl().is_err());
         });
     }
@@ -6885,7 +5220,7 @@ mod tests {
                 "-c".to_string(),
                 "print('hello')".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.pid().unwrap().is_some());
             // close drops handles then waits — this is the correct Windows lifecycle
@@ -6903,9 +5238,9 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(5)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
-            assert!(process.poll().unwrap().is_none());
+            assert!(core_pty::poll_pty_process(&process.handles, &process.returncode).unwrap().is_none());
             let _ = process.close_impl();
         });
     }
@@ -6920,7 +5255,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(0.1)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.kill_impl().is_ok());
         });
@@ -6936,7 +5271,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(0.1)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             // On Windows, terminate delegates to kill
             assert!(process.terminate_impl().is_ok());
@@ -6949,7 +5284,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             // close before start should be ok (handles are None)
             assert!(process.close_impl().is_ok());
         });
@@ -6965,7 +5300,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(0.1)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             let result = process.start_impl();
             assert!(result.is_err());
@@ -6983,7 +5318,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(2)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.resize_impl(40, 120).is_ok());
             let _ = process.close_impl();
@@ -7000,7 +5335,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(2)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             let _ = process.write_impl(b"hello\n", false);
             assert!(process.pty_input_bytes_total() >= 6);
@@ -7019,7 +5354,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(2)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert_eq!(process.pty_input_bytes_total(), 0);
             let _ = process.write_impl(b"hello\n", false);
@@ -7041,7 +5376,7 @@ mod tests {
                 "-c".to_string(),
                 "import time; time.sleep(0.1)".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             // send_interrupt on Windows writes Ctrl+C byte via PTY
             assert!(process.send_interrupt_impl().is_ok());
@@ -7057,7 +5392,7 @@ mod tests {
             let tmp = std::env::temp_dir();
             let cwd = tmp.to_str().unwrap().to_string();
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, Some(cwd), None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, Some(cwd), None, 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.close_impl().is_ok());
         });
@@ -7067,21 +5402,22 @@ mod tests {
     #[cfg(windows)]
     fn pty_process_with_env_windows() {
         pyo3::prepare_freethreaded_python();
-        pyo3::Python::with_gil(|py| {
-            let env = pyo3::types::PyDict::new(py);
+        pyo3::Python::with_gil(|_py| {
+            let mut env_pairs = Vec::new();
             if let Ok(path) = std::env::var("PATH") {
-                env.set_item("PATH", &path).unwrap();
+                env_pairs.push(("PATH".to_string(), path));
             }
             if let Ok(root) = std::env::var("SystemRoot") {
-                env.set_item("SystemRoot", &root).unwrap();
+                env_pairs.push(("SystemRoot".to_string(), root));
             }
-            env.set_item("RP_TEST_PTY", "test_value").unwrap();
+            env_pairs.push(("RP_TEST_PTY".to_string(), "test_value".to_string()));
             let argv = vec![
                 "python".to_string(),
                 "-c".to_string(),
                 "import os; print(os.environ.get('RP_TEST_PTY', 'MISSING'))".to_string(),
             ];
-            let process = NativePtyProcess::new(argv, None, Some(env), 24, 80, None).unwrap();
+            let process =
+                CoreNativePtyProcess::new(argv, None, Some(env_pairs), 24, 80, None).unwrap();
             process.start_impl().unwrap();
             assert!(process.close_impl().is_ok());
         });
@@ -7095,8 +5431,8 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
-            assert!(!process.terminal_input_relay_active());
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            assert!(!process.terminal_input_relay_active.load(Ordering::Acquire));
         });
     }
 
@@ -7106,7 +5442,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             process.stop_terminal_input_relay_impl(); // should not panic
         });
     }
@@ -7585,7 +5921,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             // reader should not be closed initially
             assert!(!process.reader.state.lock().unwrap().closed);
             process.mark_reader_closed();
@@ -7598,7 +5934,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             assert!(process.returncode.lock().unwrap().is_none());
             process.store_returncode(42);
             assert_eq!(*process.returncode.lock().unwrap(), Some(42));
@@ -7610,7 +5946,7 @@ mod tests {
         pyo3::prepare_freethreaded_python();
         pyo3::Python::with_gil(|_py| {
             let argv = vec!["python".to_string(), "-c".to_string(), "pass".to_string()];
-            let process = NativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
+            let process = CoreNativePtyProcess::new(argv, None, None, 24, 80, None).unwrap();
             assert_eq!(process.pty_input_bytes_total(), 0);
             process.record_input_metrics(b"hello\n", false);
             assert_eq!(process.pty_input_bytes_total(), 6);
@@ -7764,7 +6100,7 @@ mod tests {
     fn terminal_input_inject_multiple_events() {
         let input = NativeTerminalInput::new();
         {
-            let mut state = input.state.lock().unwrap();
+            let mut state = input.inner.state.lock().unwrap();
             for i in 0..5 {
                 state.events.push_back(TerminalInputEventRecord {
                     data: vec![b'a' + i],
@@ -7779,7 +6115,7 @@ mod tests {
         }
         assert!(input.available());
         let mut count = 0;
-        while input.next_event().is_some() {
+        while input.inner.next_event().is_some() {
             count += 1;
         }
         assert_eq!(count, 5);
@@ -7837,7 +6173,7 @@ mod tests {
         let idle = Arc::new(Mutex::new(None));
         let out_bytes = Arc::new(AtomicUsize::new(0));
         let churn_bytes = Arc::new(AtomicUsize::new(0));
-        spawn_pty_reader(
+        core_pty::spawn_pty_reader(
             reader,
             Arc::clone(&shared),
             echo,
@@ -7878,7 +6214,7 @@ mod tests {
         let idle = Arc::new(Mutex::new(None));
         let out_bytes = Arc::new(AtomicUsize::new(0));
         let churn_bytes = Arc::new(AtomicUsize::new(0));
-        spawn_pty_reader(
+        core_pty::spawn_pty_reader(
             reader,
             Arc::clone(&shared),
             echo,
@@ -7970,8 +6306,8 @@ mod tests {
     #[test]
     #[cfg(windows)]
     fn windows_pty_input_payload_via_module() {
-        assert_eq!(pty_windows::input_payload(b"hello"), b"hello");
-        assert_eq!(pty_windows::input_payload(b"\n"), b"\r");
+        assert_eq!(core_pty::windows_terminal_input_payload(b"hello"), b"hello");
+        assert_eq!(core_pty::windows_terminal_input_payload(b"\n"), b"\r");
     }
 }
