+++ src/running_process/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-__version__ = "3.0.13"
+__version__ = "3.0.14"
 
 from running_process._native import (
     ContainedProcessGroup,
